feat(dashboard): align Policy UI with mockup layout contracts
