+++ dashboard/src/policy-cloud-exceptions-ia.test.tsx
@@ -22,7 +22,7 @@ const sidebarSource = readFileSync(join(here, "approval-center-primitives.tsx"),
 
 assert(!workspaceSource.includes("PolicyExceptionForm"), "policy workspace no longer mounts PolicyExceptionForm");
 assert(!workspaceSource.includes("New exception"), "policy workspace removes local New exception copy");
-assert(tabSource.includes("Request cloud exception"), "cloud exceptions tab exposes request CTA");
+assert(tabSource.includes("+ Request cloud exception"), "cloud exceptions tab exposes request CTA");
 assert(tabSource.includes("PolicyCloudExceptionRequestPanel"), "cloud exceptions tab opens in-dashboard request panel");
 assert(tabSource.includes("fetchCloudExceptions"), "cloud exceptions tab loads synced exceptions");
 assert(tabSource.includes("PolicyCloudExceptionsSummary"), "cloud exceptions tab renders summary cards");
