+++ dashboard/src/policy-cloud-exceptions-summary.tsx
@@ -6,10 +6,11 @@ type PolicyCloudExceptionsSummaryProps = {
   loading?: boolean;
 };
 
-type SummaryTone = "blue" | "amber" | "attention" | "slate";
+type SummaryTone = "blue" | "amber" | "attention" | "slate" | "green";
 
 const SUMMARY_TONE_CLASSES: Record<SummaryTone, string> = {
   blue: "text-brand-blue",
+  green: "text-emerald-700",
   amber: "text-amber-700",
   attention: "text-brand-attention",
   slate: "text-brand-dark",
@@ -18,17 +19,20 @@ const SUMMARY_TONE_CLASSES: Record<SummaryTone, string> = {
 function SummaryCard({
   label,
   value,
+  detail,
   tone = "slate",
 }: {
   label: string;
   value: number;
+  detail: string;
   tone?: SummaryTone;
 }) {
   const toneClass = SUMMARY_TONE_CLASSES[tone];
   return (
-    <div className="rounded-xl border border-slate-200/70 bg-white p-3 text-center shadow-sm">
-      <p className={`text-2xl font-semibold tabular-nums ${toneClass}`}>{value}</p>
-      <p className="mt-1 text-[10px] font-medium uppercase tracking-wider text-muted-foreground">{label}</p>
+    <div className="rounded-xl border border-slate-200/70 bg-white p-4 shadow-sm">
+      <p className="text-[10px] font-semibold uppercase tracking-wider text-slate-500">{label}</p>
+      <p className={`mt-2 text-3xl font-semibold tabular-nums ${toneClass}`}>{value}</p>
+      <p className="mt-1 text-xs text-slate-500">{detail}</p>
     </div>
   );
 }
@@ -39,7 +43,7 @@ function SummarySkeleton() {
       {[0, 1, 2, 3].map((index) => (
         <div
           key={index}
-          className="h-[72px] animate-pulse rounded-xl border border-slate-200/70 bg-slate-100"
+          className="h-[88px] animate-pulse rounded-xl border border-slate-200/70 bg-slate-100"
           aria-hidden="true"
         />
       ))}
@@ -63,10 +67,10 @@ export function PolicyCloudExceptionsSummary({
       className="grid grid-cols-2 gap-3 md:grid-cols-4"
       aria-label="Cloud exception summary"
     >
-      <SummaryCard label="Active synced" value={activeCount} tone="blue" />
-      <SummaryCard label="Pending approval" value={pendingCount} tone="amber" />
-      <SummaryCard label="Expiring soon" value={expiringSoonCount} tone="attention" />
-      <SummaryCard label="Local ack failures" value={ackFailureCount} tone="attention" />
+      <SummaryCard label="Active synced" value={activeCount} detail="Enforced locally" tone="green" />
+      <SummaryCard label="Pending approval" value={pendingCount} detail="Awaiting decision" tone="blue" />
+      <SummaryCard label="Expiring soon" value={expiringSoonCount} detail="Within 7 days" tone="amber" />
+      <SummaryCard label="Local ack failures" value={ackFailureCount} detail="Needs attention" tone="attention" />
     </div>
   );
 }
