+++ dashboard/src/policy-workspace-page.tsx
@@ -1,6 +1,8 @@
-import { Suspense, lazy, useCallback } from "react";
+import { useCallback, useState, Suspense, lazy } from "react";
 import type { GuardPolicyDecision, GuardRuntimeSnapshot } from "./guard-types";
 import { WorkspacePageHeader } from "./workspace-page-header";
+import { PolicyPageToolbar, PolicyUnderlineTabBar } from "./policy-page-chrome";
+import type { PolicyPageView } from "./policy-workspace";
 
 const PolicyWorkspace = lazy(() =>
   import("./policy-workspace").then((module) => ({ default: module.PolicyWorkspace })),
@@ -18,9 +20,21 @@ export function PolicyWorkspacePage(props: {
   onOpenInbox: () => void;
   onRefreshPolicies: () => void;
 }) {
+  const [activeView, setActiveView] = useState<PolicyPageView>("rules");
+  const [reloading, setReloading] = useState(false);
+
   const handleOpenSettings = useCallback(() => props.onOpenSettings(), [props]);
   const handleOpenInbox = useCallback(() => props.onOpenInbox(), [props]);
-  const handleRefresh = useCallback(() => props.onRefreshPolicies(), [props]);
+  const handleViewChange = useCallback((view: PolicyPageView) => setActiveView(view), []);
+
+  const handleReloadPolicy = useCallback(() => {
+    setReloading(true);
+    try {
+      props.onRefreshPolicies();
+    } finally {
+      window.setTimeout(() => setReloading(false), 600);
+    }
+  }, [props]);
 
   return (
     <div className="space-y-6">
@@ -28,15 +42,24 @@ export function PolicyWorkspacePage(props: {
         eyebrow="Policy"
         title="Remembered rules and exceptions"
         description="See what Guard will do next time, in plain language. Remove local remembered rules here or request Cloud exceptions when Guard Cloud is connected."
+        actions={
+          <PolicyPageToolbar
+            snapshot={props.snapshot}
+            onReloadPolicy={handleReloadPolicy}
+            reloading={reloading}
+          />
+        }
       />
+      <PolicyUnderlineTabBar activeView={activeView} onViewChange={handleViewChange} />
       <Suspense fallback={<PolicyFallback />}>
         <PolicyWorkspace
+          activeView={activeView}
           policies={props.policies}
           snapshot={props.snapshot}
           onClearPolicy={props.onClearPolicy}
           onOpenSettings={handleOpenSettings}
           onOpenInbox={handleOpenInbox}
-          onRefreshPolicies={handleRefresh}
+          onOpenCloudExceptions={() => setActiveView("exceptions")}
         />
       </Suspense>
     </div>
