+++ dashboard/src/policy-workspace-views.tsx
@@ -8,6 +8,7 @@ import {
   HiMiniCube,
   HiMiniDocumentText,
   HiMiniGlobeAlt,
+  HiMiniLockClosed,
   HiMiniShieldCheck,
 } from "react-icons/hi2";
 import { Badge, Tag, EmptyState } from "./approval-center-primitives";
@@ -20,13 +21,17 @@ import {
   resolvePolicyDisplay,
   resolvePolicyEvidenceHref,
   resolvePolicyMatcherFamily,
-  resolvePolicySourceLabel,
+  resolvePolicyRowSourceLabel,
+  resolvePolicyRowTitle,
 } from "./policy-workspace-helpers";
 
 export type PolicySortKey = "app" | "updated";
 export type PolicySortState = { key: PolicySortKey; direction: "asc" | "desc" } | null;
 
 const PAGE_SIZE = 30;
+
+const RULE_GRID_CLASS =
+  "grid grid-cols-[minmax(0,1fr)] items-start gap-x-3 gap-y-3 border-b border-slate-100 px-4 py-3 last:border-0 hover:bg-slate-50/80 md:grid-cols-[72px_minmax(220px,1.4fr)_100px_96px_88px_104px_minmax(120px,1fr)]";
 
 function resolveActionTone(action: string): "success" | "destructive" | "warning" | "default" {
   if (action === "allow") {
@@ -61,83 +66,96 @@ type PolicyRuleRowProps = {
   policy: GuardPolicyDecision;
   cloudControlsUrl: string | null;
   onClear?: (policy: GuardPolicyDecision) => void;
+  cloudVariant?: boolean;
 };
 
-function PolicyRuleRow({ policy, cloudControlsUrl, onClear }: PolicyRuleRowProps) {
+function PolicyRuleRow({ policy, cloudControlsUrl, onClear, cloudVariant = false }: PolicyRuleRowProps) {
   const handleClear = useCallback(() => onClear?.(policy), [onClear, policy]);
-  const cloudManaged = isCloudManagedPolicy(policy.source);
+  const cloudManaged = cloudVariant || isCloudManagedPolicy(policy.source);
   const display = resolvePolicyDisplay(policy);
   const canClear = onClear !== undefined && !cloudManaged;
   const family = resolvePolicyMatcherFamily(policy);
   const Icon = resolveFamilyIcon(family);
+  const title = resolvePolicyRowTitle(policy, display);
+  const scopeTag = scopeLabel(policy.scope, "policy");
 
   return (
-    <tr className="border-b border-slate-100 last:border-0 hover:bg-slate-50/80">
-      <td className="w-12 px-2 py-3 align-top">
-        <div className="flex flex-col items-center gap-1.5">
-          <span className="flex h-8 w-8 shrink-0 items-center justify-center rounded-xl bg-slate-100 text-slate-500">
-            <Icon className="h-4 w-4" aria-hidden="true" />
-          </span>
-          <Badge tone={resolveActionTone(policy.action)}>{policyActionLabel(policy.action)}</Badge>
-        </div>
-      </td>
-      <td className="min-w-[220px] px-3 py-3 align-top">
-        <p className="break-all font-mono text-sm font-semibold leading-snug text-brand-dark">{display.headline}</p>
+    <article className={RULE_GRID_CLASS}>
+      <div className="flex items-start gap-2 md:col-start-1">
+        <span className="flex h-9 w-9 shrink-0 items-center justify-center rounded-xl bg-slate-100 text-slate-500">
+          <Icon className="h-4 w-4" aria-hidden="true" />
+        </span>
+        <Badge tone={resolveActionTone(policy.action)}>{policyActionLabel(policy.action)}</Badge>
+      </div>
+
+      <div className="min-w-0 md:col-start-2">
+        <p className="text-sm font-semibold leading-snug text-brand-dark">{title}</p>
         {display.kindLine ? <p className="mt-1 text-xs text-slate-500">{display.kindLine}</p> : null}
         {display.pathLine ? (
           <p className="mt-1 break-all font-mono text-[11px] leading-relaxed text-slate-500">{display.pathLine}</p>
         ) : null}
-      </td>
-      <td className="hidden min-w-[100px] px-3 py-3 align-top text-sm text-brand-dark md:table-cell">
-        {display.projectLabel ?? "—"}
-      </td>
-      <td className="hidden px-3 py-3 align-top md:table-cell">
-        <Tag tone={cloudManaged ? "blue" : "green"}>{resolvePolicySourceLabel(policy.source)}</Tag>
-      </td>
-      <td className="hidden px-3 py-3 align-top text-sm text-slate-700 lg:table-cell">
-        {scopeLabel(policy.scope, "policy")}
-      </td>
-      <td className="hidden px-3 py-3 align-top text-sm text-slate-700 xl:table-cell">
-        {harnessDisplayName(policy.harness)}
-      </td>
-      <td className="hidden whitespace-nowrap px-3 py-3 align-top text-xs text-slate-500 sm:table-cell">
-        {policy.updated_at ? formatRelativeTime(policy.updated_at) : "—"}
-      </td>
-      <td className="min-w-[120px] px-3 py-3 align-top text-right">
-        <div className="flex flex-col items-end gap-2">
-          {!cloudManaged ? (
-            <a
-              href={guardAwareHref(resolvePolicyEvidenceHref(policy))}
-              className="max-w-[160px] truncate font-mono text-xs font-medium text-brand-blue hover:underline"
-              title={resolvePolicyApprovalRecordLabel(policy)}
-            >
-              {resolvePolicyApprovalRecordLabel(policy)}
-            </a>
-          ) : null}
-          {cloudManaged && cloudControlsUrl ? (
-            <a
-              href={cloudControlsUrl}
-              target="_blank"
-              rel="noopener noreferrer"
-              className="inline-flex items-center gap-1 text-sm font-medium text-brand-blue hover:underline"
-            >
-              <HiMiniCloudArrowUp className="h-4 w-4" aria-hidden="true" />
-              View on cloud
-            </a>
-          ) : null}
-          {canClear ? (
-            <button
-              type="button"
-              onClick={handleClear}
-              className="inline-flex items-center gap-1 text-xs font-medium text-slate-500 hover:text-red-600"
-            >
-              <HiMiniTrash className="h-3.5 w-3.5" aria-hidden="true" />
-              Remove rule
-            </button>
-          ) : null}
+        <div className="mt-2 flex flex-wrap gap-3 text-xs text-slate-600 md:hidden">
+          <span>{resolvePolicyRowSourceLabel(policy)}</span>
+          <span>{scopeTag}</span>
+          <span>{harnessDisplayName(policy.harness)}</span>
         </div>
-      </td>
-    </tr>
+      </div>
+
+      <div className="hidden text-sm text-brand-dark md:col-start-3 md:block">
+        {cloudManaged ? <Tag tone="blue">{resolvePolicyRowSourceLabel(policy)}</Tag> : resolvePolicyRowSourceLabel(policy)}
+      </div>
+
+      <div className="hidden md:col-start-4 md:block">
+        <Tag tone="blue">{scopeTag}</Tag>
+      </div>
+
+      <div className="hidden text-sm text-brand-dark md:col-start-5 md:block">
+        {harnessDisplayName(policy.harness)}
+      </div>
+
+      <div className="hidden whitespace-nowrap text-xs text-slate-500 md:col-start-6 md:block">
+        {policy.updated_at ? formatRelativeTime(policy.updated_at) : "—"}
+      </div>
+
+      <div className="flex min-w-0 flex-col items-start gap-2 md:col-start-7 md:items-end">
+        {!cloudManaged ? (
+          <a
+            href={guardAwareHref(resolvePolicyEvidenceHref(policy))}
+            className="max-w-full truncate font-mono text-xs font-medium text-brand-blue hover:underline"
+            title={resolvePolicyApprovalRecordLabel(policy)}
+          >
+            {resolvePolicyApprovalRecordLabel(policy)}
+          </a>
+        ) : null}
+        {cloudManaged && cloudControlsUrl ? (
+          <a
+            href={cloudControlsUrl}
+            target="_blank"
+            rel="noopener noreferrer"
+            className="inline-flex items-center gap-1 text-sm font-medium text-brand-blue hover:underline"
+          >
+            <HiMiniCloudArrowUp className="h-4 w-4" aria-hidden="true" />
+            View on cloud
+          </a>
+        ) : null}
+        {cloudManaged ? (
+          <span className="inline-flex items-center gap-1 text-xs font-medium text-slate-500" title="Read-only Cloud policy">
+            <HiMiniLockClosed className="h-3.5 w-3.5" aria-hidden="true" />
+            Policy
+          </span>
+        ) : null}
+        {canClear ? (
+          <button
+            type="button"
+            onClick={handleClear}
+            className="inline-flex items-center gap-1 text-xs font-medium text-slate-500 hover:text-red-600"
+          >
+            <HiMiniTrash className="h-3.5 w-3.5" aria-hidden="true" />
+            Remove rule
+          </button>
+        ) : null}
+      </div>
+    </article>
   );
 }
 
@@ -147,6 +165,8 @@ type PolicyRuleTableProps = {
   onClearPolicy?: (policy: GuardPolicyDecision) => void;
   emptyTitle: string;
   emptyBody: string;
+  cloudVariant?: boolean;
+  totalCount?: number;
 };
 
 export function PolicyRuleTable({
@@ -155,11 +175,15 @@ export function PolicyRuleTable({
   onClearPolicy,
   emptyTitle,
   emptyBody,
+  cloudVariant = false,
+  totalCount,
 }: PolicyRuleTableProps) {
   const [visibleCount, setVisibleCount] = useState(PAGE_SIZE);
 
   const visiblePolicies = useMemo(() => policies.slice(0, visibleCount), [policies, visibleCount]);
-  const hasMore = policies.length > visibleCount;
+  const remaining = policies.length - visibleCount;
+  const hasMore = remaining > 0;
+  const listTotal = totalCount ?? policies.length;
 
   const handleShowMore = useCallback(() => {
     setVisibleCount((current) => current + PAGE_SIZE);
@@ -172,39 +196,38 @@ export function PolicyRuleTable({
   return (
     <div className="space-y-3">
       <div className="overflow-x-auto rounded-2xl border border-slate-100 bg-white shadow-sm">
-        <table className="min-w-[960px] w-full border-collapse text-left">
-          <thead className="border-b border-slate-100 bg-slate-50/80 text-xs font-semibold uppercase tracking-wide text-slate-500">
-            <tr>
-              <th className="px-2 py-3">Action</th>
-              <th className="px-3 py-3">Command</th>
-              <th className="hidden px-3 py-3 md:table-cell">Project</th>
-              <th className="hidden px-3 py-3 md:table-cell">Source</th>
-              <th className="hidden px-3 py-3 lg:table-cell">Scope</th>
-              <th className="hidden px-3 py-3 xl:table-cell">App</th>
-              <th className="hidden px-3 py-3 sm:table-cell">Updated</th>
-              <th className="px-3 py-3 text-right">Record</th>
-            </tr>
-          </thead>
-          <tbody>
-            {visiblePolicies.map((policy) => (
-              <PolicyRuleRow
-                key={`${policy.harness}-${policy.scope}-${policy.artifact_id ?? policy.publisher ?? "global"}-${policy.updated_at ?? ""}-${policy.source}`}
-                policy={policy}
-                cloudControlsUrl={cloudControlsUrl}
-                onClear={onClearPolicy}
-              />
-            ))}
-          </tbody>
-        </table>
+        <div
+          className="hidden border-b border-slate-100 bg-slate-50/80 px-4 py-2.5 text-[10px] font-semibold uppercase tracking-wider text-slate-500 md:grid md:grid-cols-[72px_minmax(220px,1.4fr)_100px_96px_88px_104px_minmax(120px,1fr)] md:gap-x-3"
+          aria-hidden="true"
+        >
+          <span>Action</span>
+          <span>Rule</span>
+          <span>Source</span>
+          <span>Scope</span>
+          <span>{cloudVariant ? "Applies to" : "Harness"}</span>
+          <span>Last updated</span>
+          <span className="text-right">{cloudVariant ? "Policy" : "Approval record"}</span>
+        </div>
+        <div role="list">
+          {visiblePolicies.map((policy) => (
+            <PolicyRuleRow
+              key={`${policy.harness}-${policy.scope}-${policy.artifact_id ?? policy.publisher ?? "global"}-${policy.updated_at ?? ""}-${policy.source}`}
+              policy={policy}
+              cloudControlsUrl={cloudControlsUrl}
+              onClear={onClearPolicy}
+              cloudVariant={cloudVariant}
+            />
+          ))}
+        </div>
       </div>
       {hasMore ? (
         <div className="flex justify-center pt-1">
           <button
             type="button"
             onClick={handleShowMore}
-            className="rounded-full border border-slate-200 bg-white px-4 py-2 text-sm font-medium text-brand-dark hover:bg-slate-50"
+            className="text-sm font-medium text-brand-blue hover:underline"
           >
-            Show {Math.min(PAGE_SIZE, policies.length - visibleCount)} more ({policies.length - visibleCount} remaining)
+            View all {cloudVariant ? "cloud" : "local"} rules ({listTotal}) →
           </button>
         </div>
       ) : null}
@@ -221,6 +244,7 @@ type GroupedPolicySectionProps = {
   emptyTitle: string;
   emptyBody: string;
   defaultOpen?: boolean;
+  cloudVariant?: boolean;
 };
 
 export function GroupedPolicySection({
@@ -232,9 +256,11 @@ export function GroupedPolicySection({
   emptyTitle,
   emptyBody,
   defaultOpen = true,
+  cloudVariant = false,
 }: GroupedPolicySectionProps) {
   const [open, setOpen] = useState(defaultOpen);
   const handleToggle = useCallback(() => setOpen((current) => !current), []);
+  const ruleLabel = policies.length === 1 ? "1 rule" : `${policies.length} rules`;
 
   if (policies.length === 0) {
     return (
@@ -261,7 +287,7 @@ export function GroupedPolicySection({
           <p className="text-sm text-slate-500">{description}</p>
         </div>
         <div className="flex items-center gap-2">
-          <Tag tone="slate">{policies.length}</Tag>
+          <Tag tone="slate">{ruleLabel}</Tag>
           {open ? (
             <HiMiniChevronUp className="h-4 w-4 text-slate-400" aria-hidden="true" />
           ) : (
@@ -276,6 +302,8 @@ export function GroupedPolicySection({
           onClearPolicy={onClearPolicy}
           emptyTitle={emptyTitle}
           emptyBody={emptyBody}
+          cloudVariant={cloudVariant}
+          totalCount={policies.length}
         />
       ) : null}
     </section>
