+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/feed-health-workspace.js
@@ -1,4 +1,4 @@
-import { j as jsxRuntimeExports, A as ActionButton, S as SectionLabel, w as HiMiniExclamationTriangle, aw as HiMiniArrowPath, d as HiMiniCheckCircle, bc as HiMiniSignal, I as HiMiniXCircle, aJ as HiMiniClock, m as formatRelativeTime, B as Badge, ac as Tag } from "../guard-dashboard.js";
+import { j as jsxRuntimeExports, A as ActionButton, S as SectionLabel, w as HiMiniExclamationTriangle, aw as HiMiniArrowPath, d as HiMiniCheckCircle, bj as HiMiniSignal, I as HiMiniXCircle, aT as HiMiniClock, m as formatRelativeTime, B as Badge, ac as Tag } from "../guard-dashboard.js";
 function resolveFeedSourceMode(cloudState) {
   if (cloudState === "local_only") return "sample";
   if (cloudState === "paired_waiting") return "full";
