+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/home-protection-module.js
@@ -1,4 +1,4 @@
-import { r as reactExports, j as jsxRuntimeExports, l as HiMiniShieldCheck, b7 as HiMiniInformationCircle, w as HiMiniExclamationTriangle, S as SectionLabel, A as ActionButton, b5 as HiMiniArrowRight, ac as Tag, m as formatRelativeTime, p as HiMiniChevronUp, q as HiMiniChevronDown, d as HiMiniCheckCircle, I as HiMiniXCircle } from "../guard-dashboard.js";
+import { r as reactExports, j as jsxRuntimeExports, l as HiMiniShieldCheck, bg as HiMiniInformationCircle, w as HiMiniExclamationTriangle, S as SectionLabel, A as ActionButton, bf as HiMiniArrowRight, ac as Tag, m as formatRelativeTime, p as HiMiniChevronUp, q as HiMiniChevronDown, d as HiMiniCheckCircle, I as HiMiniXCircle } from "../guard-dashboard.js";
 function resolveHomeProtectionStatus(snapshot) {
   const protection = snapshot.supply_chain?.package_manager_protection;
   if (!protection) return "unknown";
