+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/settings-workspace.js
@@ -1,6 +1,7 @@
-import { M as getDefaultExportFromCjs, r as reactExports, R as React, j as jsxRuntimeExports, v as useFocusTrap, N as HiMiniKey, S as SectionLabel, A as ActionButton, l as HiMiniShieldCheck, O as HiMiniLockClosed, Q as HiMiniBellAlert, T as HiMiniAdjustmentsHorizontal, U as HiMiniCog6Tooth, V as HiMiniCircleStack, W as TabBar, y as HiMiniChevronRight, X as resolveProtectionLevelCopy, Y as fetchSettings, Z as fetchRuntimeSnapshot, _ as updateSettings, $ as clearPolicy, a0 as clearReviewQueue, a1 as revokeApprovalGateCooldown, a2 as disableApprovalGateTotp, a3 as enrollApprovalGateTotp, a4 as verifyApprovalGateTotp, a5 as clearEvidence, a6 as exportDiagnostics, a7 as repairApprovalCenter, a8 as exportSettings, a9 as importSettings, aa as resetSettings, ab as setupDesktopNotifications, b as EmptyState, e as GuardHero, ac as Tag, ad as HiMiniMagnifyingGlass, d as HiMiniCheckCircle, w as HiMiniExclamationTriangle, ae as approvalGateCooldownLabel, o as HiMiniXMark } from "../guard-dashboard.js";
+import { M as requireReact, N as getDefaultExportFromCjs, j as jsxRuntimeExports, r as reactExports, v as useFocusTrap, O as HiMiniKey, S as SectionLabel, A as ActionButton, l as HiMiniShieldCheck, Q as HiMiniLockClosed, R as HiMiniBellAlert, T as HiMiniAdjustmentsHorizontal, U as HiMiniCog6Tooth, V as HiMiniCircleStack, W as TabBar, y as HiMiniChevronRight, X as resolveProtectionLevelCopy, Y as fetchSettings, Z as fetchRuntimeSnapshot, _ as updateSettings, $ as clearPolicy, a0 as clearReviewQueue, a1 as revokeApprovalGateCooldown, a2 as disableApprovalGateTotp, a3 as enrollApprovalGateTotp, a4 as verifyApprovalGateTotp, a5 as clearEvidence, a6 as exportDiagnostics, a7 as repairApprovalCenter, a8 as exportSettings, a9 as importSettings, aa as resetSettings, ab as setupDesktopNotifications, b as EmptyState, e as GuardHero, ac as Tag, ad as HiMiniMagnifyingGlass, d as HiMiniCheckCircle, w as HiMiniExclamationTriangle, ae as approvalGateCooldownLabel, o as HiMiniXMark } from "../guard-dashboard.js";
 import { f as filterSettingsBySearch, R as RISK_CONTROL_CONSEQUENCES, s as securityLevelLabel } from "./app-catalog.js";
-var propTypes$2 = { exports: {} };
+var lib = {};
+var propTypes = { exports: {} };
 var ReactPropTypesSecret_1;
 var hasRequiredReactPropTypesSecret;
 function requireReactPropTypesSecret() {
@@ -66,931 +67,73 @@ function requireFactoryWithThrowingShims() {
 }
 var hasRequiredPropTypes;
 function requirePropTypes() {
-  if (hasRequiredPropTypes) return propTypes$2.exports;
+  if (hasRequiredPropTypes) return propTypes.exports;
   hasRequiredPropTypes = 1;
   {
-    propTypes$2.exports = /* @__PURE__ */ requireFactoryWithThrowingShims()();
+    propTypes.exports = /* @__PURE__ */ requireFactoryWithThrowingShims()();
   }
-  return propTypes$2.exports;
+  return propTypes.exports;
 }
-var propTypesExports = /* @__PURE__ */ requirePropTypes();
-const PropTypes = /* @__PURE__ */ getDefaultExportFromCjs(propTypesExports);
-const qrcode = function(typeNumber, errorCorrectionLevel) {
-  const PAD0 = 236;
-  const PAD1 = 17;
-  let _typeNumber = typeNumber;
-  const _errorCorrectionLevel = QRErrorCorrectionLevel[errorCorrectionLevel];
-  let _modules = null;
-  let _moduleCount = 0;
-  let _dataCache = null;
-  const _dataList = [];
-  const _this = {};
-  const makeImpl = function(test, maskPattern) {
-    _moduleCount = _typeNumber * 4 + 17;
-    _modules = (function(moduleCount) {
-      const modules = new Array(moduleCount);
-      for (let row = 0; row < moduleCount; row += 1) {
-        modules[row] = new Array(moduleCount);
-        for (let col = 0; col < moduleCount; col += 1) {
-          modules[row][col] = null;
-        }
-      }
-      return modules;
-    })(_moduleCount);
-    setupPositionProbePattern(0, 0);
-    setupPositionProbePattern(_moduleCount - 7, 0);
-    setupPositionProbePattern(0, _moduleCount - 7);
-    setupPositionAdjustPattern();
-    setupTimingPattern();
-    setupTypeInfo(test, maskPattern);
-    if (_typeNumber >= 7) {
-      setupTypeNumber(test);
-    }
-    if (_dataCache == null) {
-      _dataCache = createData(_typeNumber, _errorCorrectionLevel, _dataList);
-    }
-    mapData(_dataCache, maskPattern);
+var ErrorCorrectLevel;
+var hasRequiredErrorCorrectLevel;
+function requireErrorCorrectLevel() {
+  if (hasRequiredErrorCorrectLevel) return ErrorCorrectLevel;
+  hasRequiredErrorCorrectLevel = 1;
+  ErrorCorrectLevel = {
+    L: 1,
+    M: 0,
+    Q: 3,
+    H: 2
   };
-  const setupPositionProbePattern = function(row, col) {
-    for (let r = -1; r <= 7; r += 1) {
-      if (row + r <= -1 || _moduleCount <= row + r) continue;
-      for (let c = -1; c <= 7; c += 1) {
-        if (col + c <= -1 || _moduleCount <= col + c) continue;
-        if (0 <= r && r <= 6 && (c == 0 || c == 6) || 0 <= c && c <= 6 && (r == 0 || r == 6) || 2 <= r && r <= 4 && 2 <= c && c <= 4) {
-          _modules[row + r][col + c] = true;
-        } else {
-          _modules[row + r][col + c] = false;
-        }
-      }
-    }
+  return ErrorCorrectLevel;
+}
+var mode;
+var hasRequiredMode;
+function requireMode() {
+  if (hasRequiredMode) return mode;
+  hasRequiredMode = 1;
+  mode = {
+    MODE_NUMBER: 1 << 0,
+    MODE_ALPHA_NUM: 1 << 1,
+    MODE_8BIT_BYTE: 1 << 2,
+    MODE_KANJI: 1 << 3
   };
-  const getBestMaskPattern = function() {
-    let minLostPoint = 0;
-    let pattern = 0;
-    for (let i = 0; i < 8; i += 1) {
-      makeImpl(true, i);
-      const lostPoint = QRUtil.getLostPoint(_this);
-      if (i == 0 || minLostPoint > lostPoint) {
-        minLostPoint = lostPoint;
-        pattern = i;
-      }
-    }
-    return pattern;
-  };
-  const setupTimingPattern = function() {
-    for (let r = 8; r < _moduleCount - 8; r += 1) {
-      if (_modules[r][6] != null) {
-        continue;
-      }
-      _modules[r][6] = r % 2 == 0;
-    }
-    for (let c = 8; c < _moduleCount - 8; c += 1) {
-      if (_modules[6][c] != null) {
-        continue;
-      }
-      _modules[6][c] = c % 2 == 0;
-    }
-  };
-  const setupPositionAdjustPattern = function() {
-    const pos = QRUtil.getPatternPosition(_typeNumber);
-    for (let i = 0; i < pos.length; i += 1) {
-      for (let j = 0; j < pos.length; j += 1) {
-        const row = pos[i];
-        const col = pos[j];
-        if (_modules[row][col] != null) {
-          continue;
-        }
-        for (let r = -2; r <= 2; r += 1) {
-          for (let c = -2; c <= 2; c += 1) {
-            if (r == -2 || r == 2 || c == -2 || c == 2 || r == 0 && c == 0) {
-              _modules[row + r][col + c] = true;
-            } else {
-              _modules[row + r][col + c] = false;
-            }
-          }
-        }
-      }
-    }
-  };
-  const setupTypeNumber = function(test) {
-    const bits = QRUtil.getBCHTypeNumber(_typeNumber);
-    for (let i = 0; i < 18; i += 1) {
-      const mod = !test && (bits >> i & 1) == 1;
-      _modules[Math.floor(i / 3)][i % 3 + _moduleCount - 8 - 3] = mod;
-    }
-    for (let i = 0; i < 18; i += 1) {
-      const mod = !test && (bits >> i & 1) == 1;
-      _modules[i % 3 + _moduleCount - 8 - 3][Math.floor(i / 3)] = mod;
-    }
-  };
-  const setupTypeInfo = function(test, maskPattern) {
-    const data = _errorCorrectionLevel << 3 | maskPattern;
-    const bits = QRUtil.getBCHTypeInfo(data);
-    for (let i = 0; i < 15; i += 1) {
-      const mod = !test && (bits >> i & 1) == 1;
-      if (i < 6) {
-        _modules[i][8] = mod;
-      } else if (i < 8) {
-        _modules[i + 1][8] = mod;
-      } else {
-        _modules[_moduleCount - 15 + i][8] = mod;
-      }
-    }
-    for (let i = 0; i < 15; i += 1) {
-      const mod = !test && (bits >> i & 1) == 1;
-      if (i < 8) {
-        _modules[8][_moduleCount - i - 1] = mod;
-      } else if (i < 9) {
-        _modules[8][15 - i - 1 + 1] = mod;
-      } else {
-        _modules[8][15 - i - 1] = mod;
-      }
-    }
-    _modules[_moduleCount - 8][8] = !test;
-  };
-  const mapData = function(data, maskPattern) {
-    let inc = -1;
-    let row = _moduleCount - 1;
-    let bitIndex = 7;
-    let byteIndex = 0;
-    const maskFunc = QRUtil.getMaskFunction(maskPattern);
-    for (let col = _moduleCount - 1; col > 0; col -= 2) {
-      if (col == 6) col -= 1;
-      while (true) {
-        for (let c = 0; c < 2; c += 1) {
-          if (_modules[row][col - c] == null) {
-            let dark = false;
-            if (byteIndex < data.length) {
-              dark = (data[byteIndex] >>> bitIndex & 1) == 1;
-            }
-            const mask = maskFunc(row, col - c);
-            if (mask) {
-              dark = !dark;
-            }
-            _modules[row][col - c] = dark;
-            bitIndex -= 1;
-            if (bitIndex == -1) {
-              byteIndex += 1;
-              bitIndex = 7;
-            }
-          }
-        }
-        row += inc;
-        if (row < 0 || _moduleCount <= row) {
-          row -= inc;
-          inc = -inc;
-          break;
-        }
-      }
-    }
-  };
-  const createBytes = function(buffer, rsBlocks) {
-    let offset = 0;
-    let maxDcCount = 0;
-    let maxEcCount = 0;
-    const dcdata = new Array(rsBlocks.length);
-    const ecdata = new Array(rsBlocks.length);
-    for (let r = 0; r < rsBlocks.length; r += 1) {
-      const dcCount = rsBlocks[r].dataCount;
-      const ecCount = rsBlocks[r].totalCount - dcCount;
-      maxDcCount = Math.max(maxDcCount, dcCount);
-      maxEcCount = Math.max(maxEcCount, ecCount);
-      dcdata[r] = new Array(dcCount);
-      for (let i = 0; i < dcdata[r].length; i += 1) {
-        dcdata[r][i] = 255 & buffer.getBuffer()[i + offset];
-      }
-      offset += dcCount;
-      const rsPoly = QRUtil.getErrorCorrectPolynomial(ecCount);
-      const rawPoly = qrPolynomial(dcdata[r], rsPoly.getLength() - 1);
-      const modPoly = rawPoly.mod(rsPoly);
-      ecdata[r] = new Array(rsPoly.getLength() - 1);
-      for (let i = 0; i < ecdata[r].length; i += 1) {
-        const modIndex = i + modPoly.getLength() - ecdata[r].length;
-        ecdata[r][i] = modIndex >= 0 ? modPoly.getAt(modIndex) : 0;
-      }
-    }
-    let totalCodeCount = 0;
-    for (let i = 0; i < rsBlocks.length; i += 1) {
-      totalCodeCount += rsBlocks[i].totalCount;
-    }
-    const data = new Array(totalCodeCount);
-    let index = 0;
-    for (let i = 0; i < maxDcCount; i += 1) {
-      for (let r = 0; r < rsBlocks.length; r += 1) {
-        if (i < dcdata[r].length) {
-          data[index] = dcdata[r][i];
-          index += 1;
-        }
-      }
-    }
-    for (let i = 0; i < maxEcCount; i += 1) {
-      for (let r = 0; r < rsBlocks.length; r += 1) {
-        if (i < ecdata[r].length) {
-          data[index] = ecdata[r][i];
-          index += 1;
-        }
-      }
-    }
-    return data;
-  };
-  const createData = function(typeNumber2, errorCorrectionLevel2, dataList) {
-    const rsBlocks = QRRSBlock.getRSBlocks(typeNumber2, errorCorrectionLevel2);
-    const buffer = qrBitBuffer();
-    for (let i = 0; i < dataList.length; i += 1) {
-      const data = dataList[i];
-      buffer.put(data.getMode(), 4);
-      buffer.put(data.getLength(), QRUtil.getLengthInBits(data.getMode(), typeNumber2));
-      data.write(buffer);
-    }
-    let totalDataCount = 0;
-    for (let i = 0; i < rsBlocks.length; i += 1) {
-      totalDataCount += rsBlocks[i].dataCount;
-    }
-    if (buffer.getLengthInBits() > totalDataCount * 8) {
-      throw "code length overflow. (" + buffer.getLengthInBits() + ">" + totalDataCount * 8 + ")";
-    }
-    if (buffer.getLengthInBits() + 4 <= totalDataCount * 8) {
-      buffer.put(0, 4);
-    }
-    while (buffer.getLengthInBits() % 8 != 0) {
-      buffer.putBit(false);
-    }
-    while (true) {
-      if (buffer.getLengthInBits() >= totalDataCount * 8) {
-        break;
-      }
-      buffer.put(PAD0, 8);
-      if (buffer.getLengthInBits() >= totalDataCount * 8) {
-        break;
-      }
-      buffer.put(PAD1, 8);
-    }
-    return createBytes(buffer, rsBlocks);
-  };
-  _this.addData = function(data, mode) {
-    mode = mode || "Byte";
-    let newData = null;
-    switch (mode) {
-      case "Numeric":
-        newData = qrNumber(data);
-        break;
-      case "Alphanumeric":
-        newData = qrAlphaNum(data);
-        break;
-      case "Byte":
-        newData = qr8BitByte(data);
-        break;
-      case "Kanji":
-        newData = qrKanji(data);
-        break;
-      default:
-        throw "mode:" + mode;
-    }
-    _dataList.push(newData);
-    _dataCache = null;
-  };
-  _this.isDark = function(row, col) {
-    if (row < 0 || _moduleCount <= row || col < 0 || _moduleCount <= col) {
-      throw row + "," + col;
-    }
-    return _modules[row][col];
-  };
-  _this.getModuleCount = function() {
-    return _moduleCount;
-  };
-  _this.make = function() {
-    if (_typeNumber < 1) {
-      let typeNumber2 = 1;
-      for (; typeNumber2 < 40; typeNumber2++) {
-        const rsBlocks = QRRSBlock.getRSBlocks(typeNumber2, _errorCorrectionLevel);
-        const buffer = qrBitBuffer();
-        for (let i = 0; i < _dataList.length; i++) {
-          const data = _dataList[i];
-          buffer.put(data.getMode(), 4);
-          buffer.put(data.getLength(), QRUtil.getLengthInBits(data.getMode(), typeNumber2));
-          data.write(buffer);
-        }
-        let totalDataCount = 0;
-        for (let i = 0; i < rsBlocks.length; i++) {
-          totalDataCount += rsBlocks[i].dataCount;
-        }
-        if (buffer.getLengthInBits() <= totalDataCount * 8) {
-          break;
-        }
-      }
-      _typeNumber = typeNumber2;
-    }
-    makeImpl(false, getBestMaskPattern());
-  };
-  _this.createTableTag = function(cellSize, margin) {
-    cellSize = cellSize || 2;
-    margin = typeof margin == "undefined" ? cellSize * 4 : margin;
-    let qrHtml = "";
-    qrHtml += '<table style="';
-    qrHtml += " border-width: 0px; border-style: none;";
-    qrHtml += " border-collapse: collapse;";
-    qrHtml += " padding: 0px; margin: " + margin + "px;";
-    qrHtml += '">';
-    qrHtml += "<tbody>";
-    for (let r = 0; r < _this.getModuleCount(); r += 1) {
-      qrHtml += "<tr>";
-      for (let c = 0; c < _this.getModuleCount(); c += 1) {
-        qrHtml += '<td style="';
-        qrHtml += " border-width: 0px; border-style: none;";
-        qrHtml += " border-collapse: collapse;";
-        qrHtml += " padding: 0px; margin: 0px;";
-        qrHtml += " width: " + cellSize + "px;";
-        qrHtml += " height: " + cellSize + "px;";
-        qrHtml += " background-color: ";
-        qrHtml += _this.isDark(r, c) ? "#000000" : "#ffffff";
-        qrHtml += ";";
-        qrHtml += '"/>';
-      }
-      qrHtml += "</tr>";
-    }
-    qrHtml += "</tbody>";
-    qrHtml += "</table>";
-    return qrHtml;
-  };
-  _this.createSvgTag = function(cellSize, margin, alt, title) {
-    let opts = {};
-    if (typeof arguments[0] == "object") {
-      opts = arguments[0];
-      cellSize = opts.cellSize;
-      margin = opts.margin;
-      alt = opts.alt;
-      title = opts.title;
-    }
-    cellSize = cellSize || 2;
-    margin = typeof margin == "undefined" ? cellSize * 4 : margin;
-    alt = typeof alt === "string" ? { text: alt } : alt || {};
-    alt.text = alt.text || null;
-    alt.id = alt.text ? alt.id || "qrcode-description" : null;
-    title = typeof title === "string" ? { text: title } : title || {};
-    title.text = title.text || null;
-    title.id = title.text ? title.id || "qrcode-title" : null;
-    const size = _this.getModuleCount() * cellSize + margin * 2;
-    let c, mc, r, mr, qrSvg = "", rect;
-    rect = "l" + cellSize + ",0 0," + cellSize + " -" + cellSize + ",0 0,-" + cellSize + "z ";
-    qrSvg += '<svg version="1.1" xmlns="http://www.w3.org/2000/svg"';
-    qrSvg += !opts.scalable ? ' width="' + size + 'px" height="' + size + 'px"' : "";
-    qrSvg += ' viewBox="0 0 ' + size + " " + size + '" ';
-    qrSvg += ' preserveAspectRatio="xMinYMin meet"';
-    qrSvg += title.text || alt.text ? ' role="img" aria-labelledby="' + escapeXml([title.id, alt.id].join(" ").trim()) + '"' : "";
-    qrSvg += ">";
-    qrSvg += title.text ? '<title id="' + escapeXml(title.id) + '">' + escapeXml(title.text) + "</title>" : "";
-    qrSvg += alt.text ? '<description id="' + escapeXml(alt.id) + '">' + escapeXml(alt.text) + "</description>" : "";
-    qrSvg += '<rect width="100%" height="100%" fill="white" cx="0" cy="0"/>';
-    qrSvg += '<path d="';
-    for (r = 0; r < _this.getModuleCount(); r += 1) {
-      mr = r * cellSize + margin;
-      for (c = 0; c < _this.getModuleCount(); c += 1) {
-        if (_this.isDark(r, c)) {
-          mc = c * cellSize + margin;
-          qrSvg += "M" + mc + "," + mr + rect;
-        }
-      }
-    }
-    qrSvg += '" stroke="transparent" fill="black"/>';
-    qrSvg += "</svg>";
-    return qrSvg;
-  };
-  _this.createDataURL = function(cellSize, margin) {
-    cellSize = cellSize || 2;
-    margin = typeof margin == "undefined" ? cellSize * 4 : margin;
-    const size = _this.getModuleCount() * cellSize + margin * 2;
-    const min = margin;
-    const max = size - margin;
-    return createDataURL(size, size, function(x, y) {
-      if (min <= x && x < max && min <= y && y < max) {
-        const c = Math.floor((x - min) / cellSize);
-        const r = Math.floor((y - min) / cellSize);
-        return _this.isDark(r, c) ? 0 : 1;
-      } else {
-        return 1;
-      }
-    });
-  };
-  _this.createImgTag = function(cellSize, margin, alt) {
-    cellSize = cellSize || 2;
-    margin = typeof margin == "undefined" ? cellSize * 4 : margin;
-    const size = _this.getModuleCount() * cellSize + margin * 2;
-    let img = "";
-    img += "<img";
-    img += ' src="';
-    img += _this.createDataURL(cellSize, margin);
-    img += '"';
-    img += ' width="';
-    img += size;
-    img += '"';
-    img += ' height="';
-    img += size;
-    img += '"';
-    if (alt) {
-      img += ' alt="';
-      img += escapeXml(alt);
-      img += '"';
-    }
-    img += "/>";
-    return img;
-  };
-  const escapeXml = function(s) {
-    let escaped = "";
-    for (let i = 0; i < s.length; i += 1) {
-      const c = s.charAt(i);
-      switch (c) {
-        case "<":
-          escaped += "&lt;";
-          break;
-        case ">":
-          escaped += "&gt;";
-          break;
-        case "&":
-          escaped += "&amp;";
-          break;
-        case '"':
-          escaped += "&quot;";
-          break;
-        default:
-          escaped += c;
-          break;
-      }
-    }
-    return escaped;
-  };
-  const _createHalfASCII = function(margin) {
-    const cellSize = 1;
-    margin = typeof margin == "undefined" ? cellSize * 2 : margin;
-    const size = _this.getModuleCount() * cellSize + margin * 2;
-    const min = margin;
-    const max = size - margin;
-    let y, x, r1, r2, p;
-    const blocks = {
-      "██": "█",
-      "█ ": "▀",
-      " █": "▄",
-      "  ": " "
-    };
-    const blocksLastLineNoMargin = {
-      "██": "▀",
-      "█ ": "▀",
-      " █": " ",
-      "  ": " "
-    };
-    let ascii = "";
-    for (y = 0; y < size; y += 2) {
-      r1 = Math.floor((y - min) / cellSize);
-      r2 = Math.floor((y + 1 - min) / cellSize);
-      for (x = 0; x < size; x += 1) {
-        p = "█";
-        if (min <= x && x < max && min <= y && y < max && _this.isDark(r1, Math.floor((x - min) / cellSize))) {
-          p = " ";
-        }
-        if (min <= x && x < max && min <= y + 1 && y + 1 < max && _this.isDark(r2, Math.floor((x - min) / cellSize))) {
-          p += " ";
-        } else {
-          p += "█";
-        }
-        ascii += margin < 1 && y + 1 >= max ? blocksLastLineNoMargin[p] : blocks[p];
-      }
-      ascii += "\n";
-    }
-    if (size % 2 && margin > 0) {
-      return ascii.substring(0, ascii.length - size - 1) + Array(size + 1).join("▀");
-    }
-    return ascii.substring(0, ascii.length - 1);
-  };
-  _this.createASCII = function(cellSize, margin) {
-    cellSize = cellSize || 1;
-    if (cellSize < 2) {
-      return _createHalfASCII(margin);
-    }
-    cellSize -= 1;
-    margin = typeof margin == "undefined" ? cellSize * 2 : margin;
-    const size = _this.getModuleCount() * cellSize + margin * 2;
-    const min = margin;
-    const max = size - margin;
-    let y, x, r, p;
-    const white = Array(cellSize + 1).join("██");
-    const black = Array(cellSize + 1).join("  ");
-    let ascii = "";
-    let line = "";
-    for (y = 0; y < size; y += 1) {
-      r = Math.floor((y - min) / cellSize);
-      line = "";
-      for (x = 0; x < size; x += 1) {
-        p = 1;
-        if (min <= x && x < max && min <= y && y < max && _this.isDark(r, Math.floor((x - min) / cellSize))) {
-          p = 0;
-        }
-        line += p ? white : black;
-      }
-      for (r = 0; r < cellSize; r += 1) {
-        ascii += line + "\n";
-      }
-    }
-    return ascii.substring(0, ascii.length - 1);
-  };
-  _this.renderTo2dContext = function(context, cellSize) {
-    cellSize = cellSize || 2;
-    const length = _this.getModuleCount();
-    for (let row = 0; row < length; row++) {
-      for (let col = 0; col < length; col++) {
-        context.fillStyle = _this.isDark(row, col) ? "black" : "white";
-        context.fillRect(col * cellSize, row * cellSize, cellSize, cellSize);
-      }
-    }
-  };
-  return _this;
-};
-qrcode.stringToBytes = function(s) {
-  const bytes = [];
-  for (let i = 0; i < s.length; i += 1) {
-    const c = s.charCodeAt(i);
-    bytes.push(c & 255);
+  return mode;
+}
+var _8BitByte;
+var hasRequired_8BitByte;
+function require_8BitByte() {
+  if (hasRequired_8BitByte) return _8BitByte;
+  hasRequired_8BitByte = 1;
+  var mode2 = requireMode();
+  function QR8bitByte(data) {
+    this.mode = mode2.MODE_8BIT_BYTE;
+    this.data = data;
   }
-  return bytes;
-};
-qrcode.createStringToBytes = function(unicodeData, numChars) {
-  const unicodeMap = (function() {
-    const bin = base64DecodeInputStream(unicodeData);
-    const read = function() {
-      const b = bin.read();
-      if (b == -1) throw "eof";
-      return b;
-    };
-    let count = 0;
-    const unicodeMap2 = {};
-    while (true) {
-      const b0 = bin.read();
-      if (b0 == -1) break;
-      const b1 = read();
-      const b2 = read();
-      const b3 = read();
-      const k = String.fromCharCode(b0 << 8 | b1);
-      const v = b2 << 8 | b3;
-      unicodeMap2[k] = v;
-      count += 1;
-    }
-    if (count != numChars) {
-      throw count + " != " + numChars;
-    }
-    return unicodeMap2;
-  })();
-  const unknownChar = "?".charCodeAt(0);
-  return function(s) {
-    const bytes = [];
-    for (let i = 0; i < s.length; i += 1) {
-      const c = s.charCodeAt(i);
-      if (c < 128) {
-        bytes.push(c);
-      } else {
-        const b = unicodeMap[s.charAt(i)];
-        if (typeof b == "number") {
-          if ((b & 255) == b) {
-            bytes.push(b);
-          } else {
-            bytes.push(b >>> 8);
-            bytes.push(b & 255);
-          }
-        } else {
-          bytes.push(unknownChar);
-        }
+  QR8bitByte.prototype = {
+    getLength: function(buffer) {
+      return this.data.length;
+    },
+    write: function(buffer) {
+      for (var i = 0; i < this.data.length; i++) {
+        buffer.put(this.data.charCodeAt(i), 8);
       }
     }
-    return bytes;
   };
-};
-const QRMode = {
-  MODE_NUMBER: 1 << 0,
-  MODE_ALPHA_NUM: 1 << 1,
-  MODE_8BIT_BYTE: 1 << 2,
-  MODE_KANJI: 1 << 3
-};
-const QRErrorCorrectionLevel = {
-  L: 1,
-  M: 0,
-  Q: 3,
-  H: 2
-};
-const QRMaskPattern = {
-  PATTERN000: 0,
-  PATTERN001: 1,
-  PATTERN010: 2,
-  PATTERN011: 3,
-  PATTERN100: 4,
-  PATTERN101: 5,
-  PATTERN110: 6,
-  PATTERN111: 7
-};
-const QRUtil = (function() {
-  const PATTERN_POSITION_TABLE = [
-    [],
-    [6, 18],
-    [6, 22],
-    [6, 26],
-    [6, 30],
-    [6, 34],
-    [6, 22, 38],
-    [6, 24, 42],
-    [6, 26, 46],
-    [6, 28, 50],
-    [6, 30, 54],
-    [6, 32, 58],
-    [6, 34, 62],
-    [6, 26, 46, 66],
-    [6, 26, 48, 70],
-    [6, 26, 50, 74],
-    [6, 30, 54, 78],
-    [6, 30, 56, 82],
-    [6, 30, 58, 86],
-    [6, 34, 62, 90],
-    [6, 28, 50, 72, 94],
-    [6, 26, 50, 74, 98],
-    [6, 30, 54, 78, 102],
-    [6, 28, 54, 80, 106],
-    [6, 32, 58, 84, 110],
-    [6, 30, 58, 86, 114],
-    [6, 34, 62, 90, 118],
-    [6, 26, 50, 74, 98, 122],
-    [6, 30, 54, 78, 102, 126],
-    [6, 26, 52, 78, 104, 130],
-    [6, 30, 56, 82, 108, 134],
-    [6, 34, 60, 86, 112, 138],
-    [6, 30, 58, 86, 114, 142],
-    [6, 34, 62, 90, 118, 146],
-    [6, 30, 54, 78, 102, 126, 150],
-    [6, 24, 50, 76, 102, 128, 154],
-    [6, 28, 54, 80, 106, 132, 158],
-    [6, 32, 58, 84, 110, 136, 162],
-    [6, 26, 54, 82, 110, 138, 166],
-    [6, 30, 58, 86, 114, 142, 170]
-  ];
-  const G15 = 1 << 10 | 1 << 8 | 1 << 5 | 1 << 4 | 1 << 2 | 1 << 1 | 1 << 0;
-  const G18 = 1 << 12 | 1 << 11 | 1 << 10 | 1 << 9 | 1 << 8 | 1 << 5 | 1 << 2 | 1 << 0;
-  const G15_MASK = 1 << 14 | 1 << 12 | 1 << 10 | 1 << 4 | 1 << 1;
-  const _this = {};
-  const getBCHDigit = function(data) {
-    let digit = 0;
-    while (data != 0) {
-      digit += 1;
-      data >>>= 1;
-    }
-    return digit;
-  };
-  _this.getBCHTypeInfo = function(data) {
-    let d = data << 10;
-    while (getBCHDigit(d) - getBCHDigit(G15) >= 0) {
-      d ^= G15 << getBCHDigit(d) - getBCHDigit(G15);
-    }
-    return (data << 10 | d) ^ G15_MASK;
-  };
-  _this.getBCHTypeNumber = function(data) {
-    let d = data << 12;
-    while (getBCHDigit(d) - getBCHDigit(G18) >= 0) {
-      d ^= G18 << getBCHDigit(d) - getBCHDigit(G18);
-    }
-    return data << 12 | d;
-  };
-  _this.getPatternPosition = function(typeNumber) {
-    return PATTERN_POSITION_TABLE[typeNumber - 1];
-  };
-  _this.getMaskFunction = function(maskPattern) {
-    switch (maskPattern) {
-      case QRMaskPattern.PATTERN000:
-        return function(i, j) {
-          return (i + j) % 2 == 0;
-        };
-      case QRMaskPattern.PATTERN001:
-        return function(i, j) {
-          return i % 2 == 0;
-        };
-      case QRMaskPattern.PATTERN010:
-        return function(i, j) {
-          return j % 3 == 0;
-        };
-      case QRMaskPattern.PATTERN011:
-        return function(i, j) {
-          return (i + j) % 3 == 0;
-        };
-      case QRMaskPattern.PATTERN100:
-        return function(i, j) {
-          return (Math.floor(i / 2) + Math.floor(j / 3)) % 2 == 0;
-        };
-      case QRMaskPattern.PATTERN101:
-        return function(i, j) {
-          return i * j % 2 + i * j % 3 == 0;
-        };
-      case QRMaskPattern.PATTERN110:
-        return function(i, j) {
-          return (i * j % 2 + i * j % 3) % 2 == 0;
-        };
-      case QRMaskPattern.PATTERN111:
-        return function(i, j) {
-          return (i * j % 3 + (i + j) % 2) % 2 == 0;
-        };
-      default:
-        throw "bad maskPattern:" + maskPattern;
-    }
-  };
-  _this.getErrorCorrectPolynomial = function(errorCorrectLength) {
-    let a = qrPolynomial([1], 0);
-    for (let i = 0; i < errorCorrectLength; i += 1) {
-      a = a.multiply(qrPolynomial([1, QRMath.gexp(i)], 0));
-    }
-    return a;
-  };
-  _this.getLengthInBits = function(mode, type) {
-    if (1 <= type && type < 10) {
-      switch (mode) {
-        case QRMode.MODE_NUMBER:
-          return 10;
-        case QRMode.MODE_ALPHA_NUM:
-          return 9;
-        case QRMode.MODE_8BIT_BYTE:
-          return 8;
-        case QRMode.MODE_KANJI:
-          return 8;
-        default:
-          throw "mode:" + mode;
-      }
-    } else if (type < 27) {
-      switch (mode) {
-        case QRMode.MODE_NUMBER:
-          return 12;
-        case QRMode.MODE_ALPHA_NUM:
-          return 11;
-        case QRMode.MODE_8BIT_BYTE:
-          return 16;
-        case QRMode.MODE_KANJI:
-          return 10;
-        default:
-          throw "mode:" + mode;
-      }
-    } else if (type < 41) {
-      switch (mode) {
-        case QRMode.MODE_NUMBER:
-          return 14;
-        case QRMode.MODE_ALPHA_NUM:
-          return 13;
-        case QRMode.MODE_8BIT_BYTE:
-          return 16;
-        case QRMode.MODE_KANJI:
-          return 12;
-        default:
-          throw "mode:" + mode;
-      }
-    } else {
-      throw "type:" + type;
-    }
-  };
-  _this.getLostPoint = function(qrcode2) {
-    const moduleCount = qrcode2.getModuleCount();
-    let lostPoint = 0;
-    for (let row = 0; row < moduleCount; row += 1) {
-      for (let col = 0; col < moduleCount; col += 1) {
-        let sameCount = 0;
-        const dark = qrcode2.isDark(row, col);
-        for (let r = -1; r <= 1; r += 1) {
-          if (row + r < 0 || moduleCount <= row + r) {
-            continue;
-          }
-          for (let c = -1; c <= 1; c += 1) {
-            if (col + c < 0 || moduleCount <= col + c) {
-              continue;
-            }
-            if (r == 0 && c == 0) {
-              continue;
-            }
-            if (dark == qrcode2.isDark(row + r, col + c)) {
-              sameCount += 1;
-            }
-          }
-        }
-        if (sameCount > 5) {
-          lostPoint += 3 + sameCount - 5;
-        }
-      }
-    }
-    for (let row = 0; row < moduleCount - 1; row += 1) {
-      for (let col = 0; col < moduleCount - 1; col += 1) {
-        let count = 0;
-        if (qrcode2.isDark(row, col)) count += 1;
-        if (qrcode2.isDark(row + 1, col)) count += 1;
-        if (qrcode2.isDark(row, col + 1)) count += 1;
-        if (qrcode2.isDark(row + 1, col + 1)) count += 1;
-        if (count == 0 || count == 4) {
-          lostPoint += 3;
-        }
-      }
-    }
-    for (let row = 0; row < moduleCount; row += 1) {
-      for (let col = 0; col < moduleCount - 6; col += 1) {
-        if (qrcode2.isDark(row, col) && !qrcode2.isDark(row, col + 1) && qrcode2.isDark(row, col + 2) && qrcode2.isDark(row, col + 3) && qrcode2.isDark(row, col + 4) && !qrcode2.isDark(row, col + 5) && qrcode2.isDark(row, col + 6)) {
-          lostPoint += 40;
-        }
-      }
-    }
-    for (let col = 0; col < moduleCount; col += 1) {
-      for (let row = 0; row < moduleCount - 6; row += 1) {
-        if (qrcode2.isDark(row, col) && !qrcode2.isDark(row + 1, col) && qrcode2.isDark(row + 2, col) && qrcode2.isDark(row + 3, col) && qrcode2.isDark(row + 4, col) && !qrcode2.isDark(row + 5, col) && qrcode2.isDark(row + 6, col)) {
-          lostPoint += 40;
-        }
-      }
-    }
-    let darkCount = 0;
-    for (let col = 0; col < moduleCount; col += 1) {
-      for (let row = 0; row < moduleCount; row += 1) {
-        if (qrcode2.isDark(row, col)) {
-          darkCount += 1;
-        }
-      }
-    }
-    const ratio = Math.abs(100 * darkCount / moduleCount / moduleCount - 50) / 5;
-    lostPoint += ratio * 10;
-    return lostPoint;
-  };
-  return _this;
-})();
-const QRMath = (function() {
-  const EXP_TABLE = new Array(256);
-  const LOG_TABLE = new Array(256);
-  for (let i = 0; i < 8; i += 1) {
-    EXP_TABLE[i] = 1 << i;
+  _8BitByte = QR8bitByte;
+  return _8BitByte;
+}
+var RSBlock;
+var hasRequiredRSBlock;
+function requireRSBlock() {
+  if (hasRequiredRSBlock) return RSBlock;
+  hasRequiredRSBlock = 1;
+  var ECL = requireErrorCorrectLevel();
+  function QRRSBlock(totalCount, dataCount) {
+    this.totalCount = totalCount;
+    this.dataCount = dataCount;
   }
-  for (let i = 8; i < 256; i += 1) {
-    EXP_TABLE[i] = EXP_TABLE[i - 4] ^ EXP_TABLE[i - 5] ^ EXP_TABLE[i - 6] ^ EXP_TABLE[i - 8];
-  }
-  for (let i = 0; i < 255; i += 1) {
-    LOG_TABLE[EXP_TABLE[i]] = i;
-  }
-  const _this = {};
-  _this.glog = function(n) {
-    if (n < 1) {
-      throw "glog(" + n + ")";
-    }
-    return LOG_TABLE[n];
-  };
-  _this.gexp = function(n) {
-    while (n < 0) {
-      n += 255;
-    }
-    while (n >= 256) {
-      n -= 255;
-    }
-    return EXP_TABLE[n];
-  };
-  return _this;
-})();
-const qrPolynomial = function(num, shift) {
-  if (typeof num.length == "undefined") {
-    throw num.length + "/" + shift;
-  }
-  const _num = (function() {
-    let offset = 0;
-    while (offset < num.length && num[offset] == 0) {
-      offset += 1;
-    }
-    const _num2 = new Array(num.length - offset + shift);
-    for (let i = 0; i < num.length - offset; i += 1) {
-      _num2[i] = num[i + offset];
-    }
-    return _num2;
-  })();
-  const _this = {};
-  _this.getAt = function(index) {
-    return _num[index];
-  };
-  _this.getLength = function() {
-    return _num.length;
-  };
-  _this.multiply = function(e) {
-    const num2 = new Array(_this.getLength() + e.getLength() - 1);
-    for (let i = 0; i < _this.getLength(); i += 1) {
-      for (let j = 0; j < e.getLength(); j += 1) {
-        num2[i + j] ^= QRMath.gexp(QRMath.glog(_this.getAt(i)) + QRMath.glog(e.getAt(j)));
-      }
-    }
-    return qrPolynomial(num2, 0);
-  };
-  _this.mod = function(e) {
-    if (_this.getLength() - e.getLength() < 0) {
-      return _this;
-    }
-    const ratio = QRMath.glog(_this.getAt(0)) - QRMath.glog(e.getAt(0));
-    const num2 = new Array(_this.getLength());
-    for (let i = 0; i < _this.getLength(); i += 1) {
-      num2[i] = _this.getAt(i);
-    }
-    for (let i = 0; i < e.getLength(); i += 1) {
-      num2[i] ^= QRMath.gexp(QRMath.glog(e.getAt(i)) + ratio);
-    }
-    return qrPolynomial(num2, 0).mod(e);
-  };
-  return _this;
-};
-const QRRSBlock = (function() {
-  const RS_BLOCK_TABLE = [
+  QRRSBlock.RS_BLOCK_TABLE = [
     // L
     // M
     // Q
@@ -1010,7 +153,7 @@ const QRRSBlock = (function() {
     [1, 70, 44],
     [2, 35, 17],
     [2, 35, 13],
-    // 4
+    // 4		
     [1, 100, 80],
     [2, 50, 32],
     [2, 50, 24],
@@ -1025,7 +168,7 @@ const QRRSBlock = (function() {
     [4, 43, 27],
     [4, 43, 19],
     [4, 43, 15],
-    // 7
+    // 7		
     [2, 98, 78],
     [4, 49, 31],
     [2, 32, 14, 4, 33, 15],
@@ -1040,7 +183,7 @@ const QRRSBlock = (function() {
     [3, 58, 36, 2, 59, 37],
     [4, 36, 16, 4, 37, 17],
     [4, 36, 12, 4, 37, 13],
-    // 10
+    // 10		
     [2, 86, 68, 2, 87, 69],
     [4, 69, 43, 1, 70, 44],
     [6, 43, 19, 2, 44, 20],
@@ -1069,7 +212,7 @@ const QRRSBlock = (function() {
     [5, 109, 87, 1, 110, 88],
     [5, 65, 41, 5, 66, 42],
     [5, 54, 24, 7, 55, 25],
-    [11, 36, 12, 7, 37, 13],
+    [11, 36, 12],
     // 16
     [5, 122, 98, 1, 123, 99],
     [7, 73, 45, 3, 74, 46],
@@ -1196,615 +339,869 @@ const QRRSBlock = (function() {
     [34, 54, 24, 34, 55, 25],
     [20, 45, 15, 61, 46, 16]
   ];
-  const qrRSBlock = function(totalCount, dataCount) {
-    const _this2 = {};
-    _this2.totalCount = totalCount;
-    _this2.dataCount = dataCount;
-    return _this2;
-  };
-  const _this = {};
-  const getRsBlockTable = function(typeNumber, errorCorrectionLevel) {
-    switch (errorCorrectionLevel) {
-      case QRErrorCorrectionLevel.L:
-        return RS_BLOCK_TABLE[(typeNumber - 1) * 4 + 0];
-      case QRErrorCorrectionLevel.M:
-        return RS_BLOCK_TABLE[(typeNumber - 1) * 4 + 1];
-      case QRErrorCorrectionLevel.Q:
-        return RS_BLOCK_TABLE[(typeNumber - 1) * 4 + 2];
-      case QRErrorCorrectionLevel.H:
-        return RS_BLOCK_TABLE[(typeNumber - 1) * 4 + 3];
-      default:
-        return void 0;
+  QRRSBlock.getRSBlocks = function(typeNumber, errorCorrectLevel) {
+    var rsBlock = QRRSBlock.getRsBlockTable(typeNumber, errorCorrectLevel);
+    if (rsBlock == void 0) {
+      throw new Error("bad rs block @ typeNumber:" + typeNumber + "/errorCorrectLevel:" + errorCorrectLevel);
     }
-  };
-  _this.getRSBlocks = function(typeNumber, errorCorrectionLevel) {
-    const rsBlock = getRsBlockTable(typeNumber, errorCorrectionLevel);
-    if (typeof rsBlock == "undefined") {
-      throw "bad rs block @ typeNumber:" + typeNumber + "/errorCorrectionLevel:" + errorCorrectionLevel;
-    }
-    const length = rsBlock.length / 3;
-    const list = [];
-    for (let i = 0; i < length; i += 1) {
-      const count = rsBlock[i * 3 + 0];
-      const totalCount = rsBlock[i * 3 + 1];
-      const dataCount = rsBlock[i * 3 + 2];
-      for (let j = 0; j < count; j += 1) {
-        list.push(qrRSBlock(totalCount, dataCount));
+    var length = rsBlock.length / 3;
+    var list = new Array();
+    for (var i = 0; i < length; i++) {
+      var count = rsBlock[i * 3 + 0];
+      var totalCount = rsBlock[i * 3 + 1];
+      var dataCount = rsBlock[i * 3 + 2];
+      for (var j = 0; j < count; j++) {
+        list.push(new QRRSBlock(totalCount, dataCount));
       }
     }
     return list;
   };
-  return _this;
-})();
-const qrBitBuffer = function() {
-  const _buffer = [];
-  let _length = 0;
-  const _this = {};
-  _this.getBuffer = function() {
-    return _buffer;
-  };
-  _this.getAt = function(index) {
-    const bufIndex = Math.floor(index / 8);
-    return (_buffer[bufIndex] >>> 7 - index % 8 & 1) == 1;
-  };
-  _this.put = function(num, length) {
-    for (let i = 0; i < length; i += 1) {
-      _this.putBit((num >>> length - i - 1 & 1) == 1);
+  QRRSBlock.getRsBlockTable = function(typeNumber, errorCorrectLevel) {
+    switch (errorCorrectLevel) {
+      case ECL.L:
+        return QRRSBlock.RS_BLOCK_TABLE[(typeNumber - 1) * 4 + 0];
+      case ECL.M:
+        return QRRSBlock.RS_BLOCK_TABLE[(typeNumber - 1) * 4 + 1];
+      case ECL.Q:
+        return QRRSBlock.RS_BLOCK_TABLE[(typeNumber - 1) * 4 + 2];
+      case ECL.H:
+        return QRRSBlock.RS_BLOCK_TABLE[(typeNumber - 1) * 4 + 3];
+      default:
+        return void 0;
     }
   };
-  _this.getLengthInBits = function() {
-    return _length;
-  };
-  _this.putBit = function(bit) {
-    const bufIndex = Math.floor(_length / 8);
-    if (_buffer.length <= bufIndex) {
-      _buffer.push(0);
-    }
-    if (bit) {
-      _buffer[bufIndex] |= 128 >>> _length % 8;
-    }
-    _length += 1;
-  };
-  return _this;
-};
-const qrNumber = function(data) {
-  const _mode = QRMode.MODE_NUMBER;
-  const _data = data;
-  const _this = {};
-  _this.getMode = function() {
-    return _mode;
-  };
-  _this.getLength = function(buffer) {
-    return _data.length;
-  };
-  _this.write = function(buffer) {
-    const data2 = _data;
-    let i = 0;
-    while (i + 2 < data2.length) {
-      buffer.put(strToNum(data2.substring(i, i + 3)), 10);
-      i += 3;
-    }
-    if (i < data2.length) {
-      if (data2.length - i == 1) {
-        buffer.put(strToNum(data2.substring(i, i + 1)), 4);
-      } else if (data2.length - i == 2) {
-        buffer.put(strToNum(data2.substring(i, i + 2)), 7);
+  RSBlock = QRRSBlock;
+  return RSBlock;
+}
+var BitBuffer;
+var hasRequiredBitBuffer;
+function requireBitBuffer() {
+  if (hasRequiredBitBuffer) return BitBuffer;
+  hasRequiredBitBuffer = 1;
+  function QRBitBuffer() {
+    this.buffer = new Array();
+    this.length = 0;
+  }
+  QRBitBuffer.prototype = {
+    get: function(index) {
+      var bufIndex = Math.floor(index / 8);
+      return (this.buffer[bufIndex] >>> 7 - index % 8 & 1) == 1;
+    },
+    put: function(num, length) {
+      for (var i = 0; i < length; i++) {
+        this.putBit((num >>> length - i - 1 & 1) == 1);
       }
-    }
-  };
-  const strToNum = function(s) {
-    let num = 0;
-    for (let i = 0; i < s.length; i += 1) {
-      num = num * 10 + chatToNum(s.charAt(i));
-    }
-    return num;
-  };
-  const chatToNum = function(c) {
-    if ("0" <= c && c <= "9") {
-      return c.charCodeAt(0) - "0".charCodeAt(0);
-    }
-    throw "illegal char :" + c;
-  };
-  return _this;
-};
-const qrAlphaNum = function(data) {
-  const _mode = QRMode.MODE_ALPHA_NUM;
-  const _data = data;
-  const _this = {};
-  _this.getMode = function() {
-    return _mode;
-  };
-  _this.getLength = function(buffer) {
-    return _data.length;
-  };
-  _this.write = function(buffer) {
-    const s = _data;
-    let i = 0;
-    while (i + 1 < s.length) {
-      buffer.put(
-        getCode(s.charAt(i)) * 45 + getCode(s.charAt(i + 1)),
-        11
-      );
-      i += 2;
-    }
-    if (i < s.length) {
-      buffer.put(getCode(s.charAt(i)), 6);
-    }
-  };
-  const getCode = function(c) {
-    if ("0" <= c && c <= "9") {
-      return c.charCodeAt(0) - "0".charCodeAt(0);
-    } else if ("A" <= c && c <= "Z") {
-      return c.charCodeAt(0) - "A".charCodeAt(0) + 10;
-    } else {
-      switch (c) {
-        case " ":
-          return 36;
-        case "$":
-          return 37;
-        case "%":
-          return 38;
-        case "*":
-          return 39;
-        case "+":
-          return 40;
-        case "-":
-          return 41;
-        case ".":
-          return 42;
-        case "/":
-          return 43;
-        case ":":
-          return 44;
-        default:
-          throw "illegal char :" + c;
+    },
+    getLengthInBits: function() {
+      return this.length;
+    },
+    putBit: function(bit) {
+      var bufIndex = Math.floor(this.length / 8);
+      if (this.buffer.length <= bufIndex) {
+        this.buffer.push(0);
       }
-    }
-  };
-  return _this;
-};
-const qr8BitByte = function(data) {
-  const _mode = QRMode.MODE_8BIT_BYTE;
-  const _bytes = qrcode.stringToBytes(data);
-  const _this = {};
-  _this.getMode = function() {
-    return _mode;
-  };
-  _this.getLength = function(buffer) {
-    return _bytes.length;
-  };
-  _this.write = function(buffer) {
-    for (let i = 0; i < _bytes.length; i += 1) {
-      buffer.put(_bytes[i], 8);
-    }
-  };
-  return _this;
-};
-const qrKanji = function(data) {
-  const _mode = QRMode.MODE_KANJI;
-  const stringToBytes = qrcode.stringToBytes;
-  !(function(c, code) {
-    const test = stringToBytes(c);
-    if (test.length != 2 || (test[0] << 8 | test[1]) != code) {
-      throw "sjis not supported.";
-    }
-  })("友", 38726);
-  const _bytes = stringToBytes(data);
-  const _this = {};
-  _this.getMode = function() {
-    return _mode;
-  };
-  _this.getLength = function(buffer) {
-    return ~~(_bytes.length / 2);
-  };
-  _this.write = function(buffer) {
-    const data2 = _bytes;
-    let i = 0;
-    while (i + 1 < data2.length) {
-      let c = (255 & data2[i]) << 8 | 255 & data2[i + 1];
-      if (33088 <= c && c <= 40956) {
-        c -= 33088;
-      } else if (57408 <= c && c <= 60351) {
-        c -= 49472;
-      } else {
-        throw "illegal char at " + (i + 1) + "/" + c;
+      if (bit) {
+        this.buffer[bufIndex] |= 128 >>> this.length % 8;
       }
-      c = (c >>> 8 & 255) * 192 + (c & 255);
-      buffer.put(c, 13);
-      i += 2;
-    }
-    if (i < data2.length) {
-      throw "illegal char at " + (i + 1);
+      this.length++;
     }
   };
-  return _this;
-};
-const byteArrayOutputStream = function() {
-  const _bytes = [];
-  const _this = {};
-  _this.writeByte = function(b) {
-    _bytes.push(b & 255);
-  };
-  _this.writeShort = function(i) {
-    _this.writeByte(i);
-    _this.writeByte(i >>> 8);
-  };
-  _this.writeBytes = function(b, off, len) {
-    off = off || 0;
-    len = len || b.length;
-    for (let i = 0; i < len; i += 1) {
-      _this.writeByte(b[i + off]);
-    }
-  };
-  _this.writeString = function(s) {
-    for (let i = 0; i < s.length; i += 1) {
-      _this.writeByte(s.charCodeAt(i));
-    }
-  };
-  _this.toByteArray = function() {
-    return _bytes;
-  };
-  _this.toString = function() {
-    let s = "";
-    s += "[";
-    for (let i = 0; i < _bytes.length; i += 1) {
-      if (i > 0) {
-        s += ",";
+  BitBuffer = QRBitBuffer;
+  return BitBuffer;
+}
+var math;
+var hasRequiredMath;
+function requireMath() {
+  if (hasRequiredMath) return math;
+  hasRequiredMath = 1;
+  var QRMath = {
+    glog: function(n) {
+      if (n < 1) {
+        throw new Error("glog(" + n + ")");
       }
-      s += _bytes[i];
-    }
-    s += "]";
-    return s;
-  };
-  return _this;
-};
-const base64EncodeOutputStream = function() {
-  let _buffer = 0;
-  let _buflen = 0;
-  let _length = 0;
-  let _base64 = "";
-  const _this = {};
-  const writeEncoded = function(b) {
-    _base64 += String.fromCharCode(encode(b & 63));
-  };
-  const encode = function(n) {
-    if (n < 0) {
-      throw "n:" + n;
-    } else if (n < 26) {
-      return 65 + n;
-    } else if (n < 52) {
-      return 97 + (n - 26);
-    } else if (n < 62) {
-      return 48 + (n - 52);
-    } else if (n == 62) {
-      return 43;
-    } else if (n == 63) {
-      return 47;
-    } else {
-      throw "n:" + n;
-    }
-  };
-  _this.writeByte = function(n) {
-    _buffer = _buffer << 8 | n & 255;
-    _buflen += 8;
-    _length += 1;
-    while (_buflen >= 6) {
-      writeEncoded(_buffer >>> _buflen - 6);
-      _buflen -= 6;
-    }
-  };
-  _this.flush = function() {
-    if (_buflen > 0) {
-      writeEncoded(_buffer << 6 - _buflen);
-      _buffer = 0;
-      _buflen = 0;
-    }
-    if (_length % 3 != 0) {
-      const padlen = 3 - _length % 3;
-      for (let i = 0; i < padlen; i += 1) {
-        _base64 += "=";
+      return QRMath.LOG_TABLE[n];
+    },
+    gexp: function(n) {
+      while (n < 0) {
+        n += 255;
       }
+      while (n >= 256) {
+        n -= 255;
+      }
+      return QRMath.EXP_TABLE[n];
+    },
+    EXP_TABLE: new Array(256),
+    LOG_TABLE: new Array(256)
+  };
+  for (var i = 0; i < 8; i++) {
+    QRMath.EXP_TABLE[i] = 1 << i;
+  }
+  for (var i = 8; i < 256; i++) {
+    QRMath.EXP_TABLE[i] = QRMath.EXP_TABLE[i - 4] ^ QRMath.EXP_TABLE[i - 5] ^ QRMath.EXP_TABLE[i - 6] ^ QRMath.EXP_TABLE[i - 8];
+  }
+  for (var i = 0; i < 255; i++) {
+    QRMath.LOG_TABLE[QRMath.EXP_TABLE[i]] = i;
+  }
+  math = QRMath;
+  return math;
+}
+var Polynomial;
+var hasRequiredPolynomial;
+function requirePolynomial() {
+  if (hasRequiredPolynomial) return Polynomial;
+  hasRequiredPolynomial = 1;
+  var math2 = requireMath();
+  function QRPolynomial(num, shift) {
+    if (num.length == void 0) {
+      throw new Error(num.length + "/" + shift);
     }
-  };
-  _this.toString = function() {
-    return _base64;
-  };
-  return _this;
-};
-const base64DecodeInputStream = function(str) {
-  const _str = str;
-  let _pos = 0;
-  let _buffer = 0;
-  let _buflen = 0;
-  const _this = {};
-  _this.read = function() {
-    while (_buflen < 8) {
-      if (_pos >= _str.length) {
-        if (_buflen == 0) {
-          return -1;
+    var offset = 0;
+    while (offset < num.length && num[offset] == 0) {
+      offset++;
+    }
+    this.num = new Array(num.length - offset + shift);
+    for (var i = 0; i < num.length - offset; i++) {
+      this.num[i] = num[i + offset];
+    }
+  }
+  QRPolynomial.prototype = {
+    get: function(index) {
+      return this.num[index];
+    },
+    getLength: function() {
+      return this.num.length;
+    },
+    multiply: function(e) {
+      var num = new Array(this.getLength() + e.getLength() - 1);
+      for (var i = 0; i < this.getLength(); i++) {
+        for (var j = 0; j < e.getLength(); j++) {
+          num[i + j] ^= math2.gexp(math2.glog(this.get(i)) + math2.glog(e.get(j)));
         }
-        throw "unexpected end of file./" + _buflen;
       }
-      const c = _str.charAt(_pos);
-      _pos += 1;
-      if (c == "=") {
-        _buflen = 0;
-        return -1;
-      } else if (c.match(/^\s$/)) {
+      return new QRPolynomial(num, 0);
+    },
+    mod: function(e) {
+      if (this.getLength() - e.getLength() < 0) {
+        return this;
+      }
+      var ratio = math2.glog(this.get(0)) - math2.glog(e.get(0));
+      var num = new Array(this.getLength());
+      for (var i = 0; i < this.getLength(); i++) {
+        num[i] = this.get(i);
+      }
+      for (var i = 0; i < e.getLength(); i++) {
+        num[i] ^= math2.gexp(math2.glog(e.get(i)) + ratio);
+      }
+      return new QRPolynomial(num, 0).mod(e);
+    }
+  };
+  Polynomial = QRPolynomial;
+  return Polynomial;
+}
+var util;
+var hasRequiredUtil;
+function requireUtil() {
+  if (hasRequiredUtil) return util;
+  hasRequiredUtil = 1;
+  var Mode = requireMode();
+  var Polynomial2 = requirePolynomial();
+  var math2 = requireMath();
+  var QRMaskPattern = {
+    PATTERN000: 0,
+    PATTERN001: 1,
+    PATTERN010: 2,
+    PATTERN011: 3,
+    PATTERN100: 4,
+    PATTERN101: 5,
+    PATTERN110: 6,
+    PATTERN111: 7
+  };
+  var QRUtil = {
+    PATTERN_POSITION_TABLE: [
+      [],
+      [6, 18],
+      [6, 22],
+      [6, 26],
+      [6, 30],
+      [6, 34],
+      [6, 22, 38],
+      [6, 24, 42],
+      [6, 26, 46],
+      [6, 28, 50],
+      [6, 30, 54],
+      [6, 32, 58],
+      [6, 34, 62],
+      [6, 26, 46, 66],
+      [6, 26, 48, 70],
+      [6, 26, 50, 74],
+      [6, 30, 54, 78],
+      [6, 30, 56, 82],
+      [6, 30, 58, 86],
+      [6, 34, 62, 90],
+      [6, 28, 50, 72, 94],
+      [6, 26, 50, 74, 98],
+      [6, 30, 54, 78, 102],
+      [6, 28, 54, 80, 106],
+      [6, 32, 58, 84, 110],
+      [6, 30, 58, 86, 114],
+      [6, 34, 62, 90, 118],
+      [6, 26, 50, 74, 98, 122],
+      [6, 30, 54, 78, 102, 126],
+      [6, 26, 52, 78, 104, 130],
+      [6, 30, 56, 82, 108, 134],
+      [6, 34, 60, 86, 112, 138],
+      [6, 30, 58, 86, 114, 142],
+      [6, 34, 62, 90, 118, 146],
+      [6, 30, 54, 78, 102, 126, 150],
+      [6, 24, 50, 76, 102, 128, 154],
+      [6, 28, 54, 80, 106, 132, 158],
+      [6, 32, 58, 84, 110, 136, 162],
+      [6, 26, 54, 82, 110, 138, 166],
+      [6, 30, 58, 86, 114, 142, 170]
+    ],
+    G15: 1 << 10 | 1 << 8 | 1 << 5 | 1 << 4 | 1 << 2 | 1 << 1 | 1 << 0,
+    G18: 1 << 12 | 1 << 11 | 1 << 10 | 1 << 9 | 1 << 8 | 1 << 5 | 1 << 2 | 1 << 0,
+    G15_MASK: 1 << 14 | 1 << 12 | 1 << 10 | 1 << 4 | 1 << 1,
+    getBCHTypeInfo: function(data) {
+      var d = data << 10;
+      while (QRUtil.getBCHDigit(d) - QRUtil.getBCHDigit(QRUtil.G15) >= 0) {
+        d ^= QRUtil.G15 << QRUtil.getBCHDigit(d) - QRUtil.getBCHDigit(QRUtil.G15);
+      }
+      return (data << 10 | d) ^ QRUtil.G15_MASK;
+    },
+    getBCHTypeNumber: function(data) {
+      var d = data << 12;
+      while (QRUtil.getBCHDigit(d) - QRUtil.getBCHDigit(QRUtil.G18) >= 0) {
+        d ^= QRUtil.G18 << QRUtil.getBCHDigit(d) - QRUtil.getBCHDigit(QRUtil.G18);
+      }
+      return data << 12 | d;
+    },
+    getBCHDigit: function(data) {
+      var digit = 0;
+      while (data != 0) {
+        digit++;
+        data >>>= 1;
+      }
+      return digit;
+    },
+    getPatternPosition: function(typeNumber) {
+      return QRUtil.PATTERN_POSITION_TABLE[typeNumber - 1];
+    },
+    getMask: function(maskPattern, i, j) {
+      switch (maskPattern) {
+        case QRMaskPattern.PATTERN000:
+          return (i + j) % 2 == 0;
+        case QRMaskPattern.PATTERN001:
+          return i % 2 == 0;
+        case QRMaskPattern.PATTERN010:
+          return j % 3 == 0;
+        case QRMaskPattern.PATTERN011:
+          return (i + j) % 3 == 0;
+        case QRMaskPattern.PATTERN100:
+          return (Math.floor(i / 2) + Math.floor(j / 3)) % 2 == 0;
+        case QRMaskPattern.PATTERN101:
+          return i * j % 2 + i * j % 3 == 0;
+        case QRMaskPattern.PATTERN110:
+          return (i * j % 2 + i * j % 3) % 2 == 0;
+        case QRMaskPattern.PATTERN111:
+          return (i * j % 3 + (i + j) % 2) % 2 == 0;
+        default:
+          throw new Error("bad maskPattern:" + maskPattern);
+      }
+    },
+    getErrorCorrectPolynomial: function(errorCorrectLength) {
+      var a = new Polynomial2([1], 0);
+      for (var i = 0; i < errorCorrectLength; i++) {
+        a = a.multiply(new Polynomial2([1, math2.gexp(i)], 0));
+      }
+      return a;
+    },
+    getLengthInBits: function(mode2, type) {
+      if (1 <= type && type < 10) {
+        switch (mode2) {
+          case Mode.MODE_NUMBER:
+            return 10;
+          case Mode.MODE_ALPHA_NUM:
+            return 9;
+          case Mode.MODE_8BIT_BYTE:
+            return 8;
+          case Mode.MODE_KANJI:
+            return 8;
+          default:
+            throw new Error("mode:" + mode2);
+        }
+      } else if (type < 27) {
+        switch (mode2) {
+          case Mode.MODE_NUMBER:
+            return 12;
+          case Mode.MODE_ALPHA_NUM:
+            return 11;
+          case Mode.MODE_8BIT_BYTE:
+            return 16;
+          case Mode.MODE_KANJI:
+            return 10;
+          default:
+            throw new Error("mode:" + mode2);
+        }
+      } else if (type < 41) {
+        switch (mode2) {
+          case Mode.MODE_NUMBER:
+            return 14;
+          case Mode.MODE_ALPHA_NUM:
+            return 13;
+          case Mode.MODE_8BIT_BYTE:
+            return 16;
+          case Mode.MODE_KANJI:
+            return 12;
+          default:
+            throw new Error("mode:" + mode2);
+        }
+      } else {
+        throw new Error("type:" + type);
+      }
+    },
+    getLostPoint: function(qrCode) {
+      var moduleCount = qrCode.getModuleCount();
+      var lostPoint = 0;
+      for (var row = 0; row < moduleCount; row++) {
+        for (var col = 0; col < moduleCount; col++) {
+          var sameCount = 0;
+          var dark = qrCode.isDark(row, col);
+          for (var r = -1; r <= 1; r++) {
+            if (row + r < 0 || moduleCount <= row + r) {
+              continue;
+            }
+            for (var c = -1; c <= 1; c++) {
+              if (col + c < 0 || moduleCount <= col + c) {
+                continue;
+              }
+              if (r == 0 && c == 0) {
+                continue;
+              }
+              if (dark == qrCode.isDark(row + r, col + c)) {
+                sameCount++;
+              }
+            }
+          }
+          if (sameCount > 5) {
+            lostPoint += 3 + sameCount - 5;
+          }
+        }
+      }
+      for (var row = 0; row < moduleCount - 1; row++) {
+        for (var col = 0; col < moduleCount - 1; col++) {
+          var count = 0;
+          if (qrCode.isDark(row, col)) count++;
+          if (qrCode.isDark(row + 1, col)) count++;
+          if (qrCode.isDark(row, col + 1)) count++;
+          if (qrCode.isDark(row + 1, col + 1)) count++;
+          if (count == 0 || count == 4) {
+            lostPoint += 3;
+          }
+        }
+      }
+      for (var row = 0; row < moduleCount; row++) {
+        for (var col = 0; col < moduleCount - 6; col++) {
+          if (qrCode.isDark(row, col) && !qrCode.isDark(row, col + 1) && qrCode.isDark(row, col + 2) && qrCode.isDark(row, col + 3) && qrCode.isDark(row, col + 4) && !qrCode.isDark(row, col + 5) && qrCode.isDark(row, col + 6)) {
+            lostPoint += 40;
+          }
+        }
+      }
+      for (var col = 0; col < moduleCount; col++) {
+        for (var row = 0; row < moduleCount - 6; row++) {
+          if (qrCode.isDark(row, col) && !qrCode.isDark(row + 1, col) && qrCode.isDark(row + 2, col) && qrCode.isDark(row + 3, col) && qrCode.isDark(row + 4, col) && !qrCode.isDark(row + 5, col) && qrCode.isDark(row + 6, col)) {
+            lostPoint += 40;
+          }
+        }
+      }
+      var darkCount = 0;
+      for (var col = 0; col < moduleCount; col++) {
+        for (var row = 0; row < moduleCount; row++) {
+          if (qrCode.isDark(row, col)) {
+            darkCount++;
+          }
+        }
+      }
+      var ratio = Math.abs(100 * darkCount / moduleCount / moduleCount - 50) / 5;
+      lostPoint += ratio * 10;
+      return lostPoint;
+    }
+  };
+  util = QRUtil;
+  return util;
+}
+var QRCode_1;
+var hasRequiredQRCode;
+function requireQRCode() {
+  if (hasRequiredQRCode) return QRCode_1;
+  hasRequiredQRCode = 1;
+  var BitByte = require_8BitByte();
+  var RSBlock2 = requireRSBlock();
+  var BitBuffer2 = requireBitBuffer();
+  var util2 = requireUtil();
+  var Polynomial2 = requirePolynomial();
+  function QRCode2(typeNumber, errorCorrectLevel) {
+    this.typeNumber = typeNumber;
+    this.errorCorrectLevel = errorCorrectLevel;
+    this.modules = null;
+    this.moduleCount = 0;
+    this.dataCache = null;
+    this.dataList = [];
+  }
+  var proto = QRCode2.prototype;
+  proto.addData = function(data) {
+    var newData = new BitByte(data);
+    this.dataList.push(newData);
+    this.dataCache = null;
+  };
+  proto.isDark = function(row, col) {
+    if (row < 0 || this.moduleCount <= row || col < 0 || this.moduleCount <= col) {
+      throw new Error(row + "," + col);
+    }
+    return this.modules[row][col];
+  };
+  proto.getModuleCount = function() {
+    return this.moduleCount;
+  };
+  proto.make = function() {
+    if (this.typeNumber < 1) {
+      var typeNumber = 1;
+      for (typeNumber = 1; typeNumber < 40; typeNumber++) {
+        var rsBlocks = RSBlock2.getRSBlocks(typeNumber, this.errorCorrectLevel);
+        var buffer = new BitBuffer2();
+        var totalDataCount = 0;
+        for (var i = 0; i < rsBlocks.length; i++) {
+          totalDataCount += rsBlocks[i].dataCount;
+        }
+        for (var i = 0; i < this.dataList.length; i++) {
+          var data = this.dataList[i];
+          buffer.put(data.mode, 4);
+          buffer.put(data.getLength(), util2.getLengthInBits(data.mode, typeNumber));
+          data.write(buffer);
+        }
+        if (buffer.getLengthInBits() <= totalDataCount * 8)
+          break;
+      }
+      this.typeNumber = typeNumber;
+    }
+    this.makeImpl(false, this.getBestMaskPattern());
+  };
+  proto.makeImpl = function(test, maskPattern) {
+    this.moduleCount = this.typeNumber * 4 + 17;
+    this.modules = new Array(this.moduleCount);
+    for (var row = 0; row < this.moduleCount; row++) {
+      this.modules[row] = new Array(this.moduleCount);
+      for (var col = 0; col < this.moduleCount; col++) {
+        this.modules[row][col] = null;
+      }
+    }
+    this.setupPositionProbePattern(0, 0);
+    this.setupPositionProbePattern(this.moduleCount - 7, 0);
+    this.setupPositionProbePattern(0, this.moduleCount - 7);
+    this.setupPositionAdjustPattern();
+    this.setupTimingPattern();
+    this.setupTypeInfo(test, maskPattern);
+    if (this.typeNumber >= 7) {
+      this.setupTypeNumber(test);
+    }
+    if (this.dataCache == null) {
+      this.dataCache = QRCode2.createData(this.typeNumber, this.errorCorrectLevel, this.dataList);
+    }
+    this.mapData(this.dataCache, maskPattern);
+  };
+  proto.setupPositionProbePattern = function(row, col) {
+    for (var r = -1; r <= 7; r++) {
+      if (row + r <= -1 || this.moduleCount <= row + r) continue;
+      for (var c = -1; c <= 7; c++) {
+        if (col + c <= -1 || this.moduleCount <= col + c) continue;
+        if (0 <= r && r <= 6 && (c == 0 || c == 6) || 0 <= c && c <= 6 && (r == 0 || r == 6) || 2 <= r && r <= 4 && 2 <= c && c <= 4) {
+          this.modules[row + r][col + c] = true;
+        } else {
+          this.modules[row + r][col + c] = false;
+        }
+      }
+    }
+  };
+  proto.getBestMaskPattern = function() {
+    var minLostPoint = 0;
+    var pattern = 0;
+    for (var i = 0; i < 8; i++) {
+      this.makeImpl(true, i);
+      var lostPoint = util2.getLostPoint(this);
+      if (i == 0 || minLostPoint > lostPoint) {
+        minLostPoint = lostPoint;
+        pattern = i;
+      }
+    }
+    return pattern;
+  };
+  proto.createMovieClip = function(target_mc, instance_name, depth) {
+    var qr_mc = target_mc.createEmptyMovieClip(instance_name, depth);
+    var cs = 1;
+    this.make();
+    for (var row = 0; row < this.modules.length; row++) {
+      var y = row * cs;
+      for (var col = 0; col < this.modules[row].length; col++) {
+        var x = col * cs;
+        var dark = this.modules[row][col];
+        if (dark) {
+          qr_mc.beginFill(0, 100);
+          qr_mc.moveTo(x, y);
+          qr_mc.lineTo(x + cs, y);
+          qr_mc.lineTo(x + cs, y + cs);
+          qr_mc.lineTo(x, y + cs);
+          qr_mc.endFill();
+        }
+      }
+    }
+    return qr_mc;
+  };
+  proto.setupTimingPattern = function() {
+    for (var r = 8; r < this.moduleCount - 8; r++) {
+      if (this.modules[r][6] != null) {
         continue;
       }
-      _buffer = _buffer << 6 | decode(c.charCodeAt(0));
-      _buflen += 6;
+      this.modules[r][6] = r % 2 == 0;
     }
-    const n = _buffer >>> _buflen - 8 & 255;
-    _buflen -= 8;
-    return n;
-  };
-  const decode = function(c) {
-    if (65 <= c && c <= 90) {
-      return c - 65;
-    } else if (97 <= c && c <= 122) {
-      return c - 97 + 26;
-    } else if (48 <= c && c <= 57) {
-      return c - 48 + 52;
-    } else if (c == 43) {
-      return 62;
-    } else if (c == 47) {
-      return 63;
-    } else {
-      throw "c:" + c;
-    }
-  };
-  return _this;
-};
-const gifImage = function(width, height) {
-  const _width = width;
-  const _height = height;
-  const _data = new Array(width * height);
-  const _this = {};
-  _this.setPixel = function(x, y, pixel) {
-    _data[y * _width + x] = pixel;
-  };
-  _this.write = function(out) {
-    out.writeString("GIF87a");
-    out.writeShort(_width);
-    out.writeShort(_height);
-    out.writeByte(128);
-    out.writeByte(0);
-    out.writeByte(0);
-    out.writeByte(0);
-    out.writeByte(0);
-    out.writeByte(0);
-    out.writeByte(255);
-    out.writeByte(255);
-    out.writeByte(255);
-    out.writeString(",");
-    out.writeShort(0);
-    out.writeShort(0);
-    out.writeShort(_width);
-    out.writeShort(_height);
-    out.writeByte(0);
-    const lzwMinCodeSize = 2;
-    const raster = getLZWRaster(lzwMinCodeSize);
-    out.writeByte(lzwMinCodeSize);
-    let offset = 0;
-    while (raster.length - offset > 255) {
-      out.writeByte(255);
-      out.writeBytes(raster, offset, 255);
-      offset += 255;
-    }
-    out.writeByte(raster.length - offset);
-    out.writeBytes(raster, offset, raster.length - offset);
-    out.writeByte(0);
-    out.writeString(";");
-  };
-  const bitOutputStream = function(out) {
-    const _out = out;
-    let _bitLength = 0;
-    let _bitBuffer = 0;
-    const _this2 = {};
-    _this2.write = function(data, length) {
-      if (data >>> length != 0) {
-        throw "length over";
+    for (var c = 8; c < this.moduleCount - 8; c++) {
+      if (this.modules[6][c] != null) {
+        continue;
       }
-      while (_bitLength + length >= 8) {
-        _out.writeByte(255 & (data << _bitLength | _bitBuffer));
-        length -= 8 - _bitLength;
-        data >>>= 8 - _bitLength;
-        _bitBuffer = 0;
-        _bitLength = 0;
-      }
-      _bitBuffer = data << _bitLength | _bitBuffer;
-      _bitLength = _bitLength + length;
-    };
-    _this2.flush = function() {
-      if (_bitLength > 0) {
-        _out.writeByte(_bitBuffer);
-      }
-    };
-    return _this2;
-  };
-  const getLZWRaster = function(lzwMinCodeSize) {
-    const clearCode = 1 << lzwMinCodeSize;
-    const endCode = (1 << lzwMinCodeSize) + 1;
-    let bitLength = lzwMinCodeSize + 1;
-    const table = lzwTable();
-    for (let i = 0; i < clearCode; i += 1) {
-      table.add(String.fromCharCode(i));
+      this.modules[6][c] = c % 2 == 0;
     }
-    table.add(String.fromCharCode(clearCode));
-    table.add(String.fromCharCode(endCode));
-    const byteOut = byteArrayOutputStream();
-    const bitOut = bitOutputStream(byteOut);
-    bitOut.write(clearCode, bitLength);
-    let dataIndex = 0;
-    let s = String.fromCharCode(_data[dataIndex]);
-    dataIndex += 1;
-    while (dataIndex < _data.length) {
-      const c = String.fromCharCode(_data[dataIndex]);
-      dataIndex += 1;
-      if (table.contains(s + c)) {
-        s = s + c;
-      } else {
-        bitOut.write(table.indexOf(s), bitLength);
-        if (table.size() < 4095) {
-          if (table.size() == 1 << bitLength) {
-            bitLength += 1;
-          }
-          table.add(s + c);
+  };
+  proto.setupPositionAdjustPattern = function() {
+    var pos = util2.getPatternPosition(this.typeNumber);
+    for (var i = 0; i < pos.length; i++) {
+      for (var j = 0; j < pos.length; j++) {
+        var row = pos[i];
+        var col = pos[j];
+        if (this.modules[row][col] != null) {
+          continue;
         }
-        s = c;
+        for (var r = -2; r <= 2; r++) {
+          for (var c = -2; c <= 2; c++) {
+            if (r == -2 || r == 2 || c == -2 || c == 2 || r == 0 && c == 0) {
+              this.modules[row + r][col + c] = true;
+            } else {
+              this.modules[row + r][col + c] = false;
+            }
+          }
+        }
       }
     }
-    bitOut.write(table.indexOf(s), bitLength);
-    bitOut.write(endCode, bitLength);
-    bitOut.flush();
-    return byteOut.toByteArray();
   };
-  const lzwTable = function() {
-    const _map = {};
-    let _size = 0;
-    const _this2 = {};
-    _this2.add = function(key) {
-      if (_this2.contains(key)) {
-        throw "dup key:" + key;
+  proto.setupTypeNumber = function(test) {
+    var bits = util2.getBCHTypeNumber(this.typeNumber);
+    for (var i = 0; i < 18; i++) {
+      var mod = !test && (bits >> i & 1) == 1;
+      this.modules[Math.floor(i / 3)][i % 3 + this.moduleCount - 8 - 3] = mod;
+    }
+    for (var i = 0; i < 18; i++) {
+      var mod = !test && (bits >> i & 1) == 1;
+      this.modules[i % 3 + this.moduleCount - 8 - 3][Math.floor(i / 3)] = mod;
+    }
+  };
+  proto.setupTypeInfo = function(test, maskPattern) {
+    var data = this.errorCorrectLevel << 3 | maskPattern;
+    var bits = util2.getBCHTypeInfo(data);
+    for (var i = 0; i < 15; i++) {
+      var mod = !test && (bits >> i & 1) == 1;
+      if (i < 6) {
+        this.modules[i][8] = mod;
+      } else if (i < 8) {
+        this.modules[i + 1][8] = mod;
+      } else {
+        this.modules[this.moduleCount - 15 + i][8] = mod;
       }
-      _map[key] = _size;
-      _size += 1;
-    };
-    _this2.size = function() {
-      return _size;
-    };
-    _this2.indexOf = function(key) {
-      return _map[key];
-    };
-    _this2.contains = function(key) {
-      return typeof _map[key] != "undefined";
-    };
-    return _this2;
+    }
+    for (var i = 0; i < 15; i++) {
+      var mod = !test && (bits >> i & 1) == 1;
+      if (i < 8) {
+        this.modules[8][this.moduleCount - i - 1] = mod;
+      } else if (i < 9) {
+        this.modules[8][15 - i - 1 + 1] = mod;
+      } else {
+        this.modules[8][15 - i - 1] = mod;
+      }
+    }
+    this.modules[this.moduleCount - 8][8] = !test;
   };
-  return _this;
-};
-const createDataURL = function(width, height, getPixel) {
-  const gif = gifImage(width, height);
-  for (let y = 0; y < height; y += 1) {
-    for (let x = 0; x < width; x += 1) {
-      gif.setPixel(x, y, getPixel(x, y));
+  proto.mapData = function(data, maskPattern) {
+    var inc = -1;
+    var row = this.moduleCount - 1;
+    var bitIndex = 7;
+    var byteIndex = 0;
+    for (var col = this.moduleCount - 1; col > 0; col -= 2) {
+      if (col == 6) col--;
+      while (true) {
+        for (var c = 0; c < 2; c++) {
+          if (this.modules[row][col - c] == null) {
+            var dark = false;
+            if (byteIndex < data.length) {
+              dark = (data[byteIndex] >>> bitIndex & 1) == 1;
+            }
+            var mask = util2.getMask(maskPattern, row, col - c);
+            if (mask) {
+              dark = !dark;
+            }
+            this.modules[row][col - c] = dark;
+            bitIndex--;
+            if (bitIndex == -1) {
+              byteIndex++;
+              bitIndex = 7;
+            }
+          }
+        }
+        row += inc;
+        if (row < 0 || this.moduleCount <= row) {
+          row -= inc;
+          inc = -inc;
+          break;
+        }
+      }
     }
-  }
-  const b = byteArrayOutputStream();
-  gif.write(b);
-  const base64 = base64EncodeOutputStream();
-  const bytes = b.toByteArray();
-  for (let i = 0; i < bytes.length; i += 1) {
-    base64.writeByte(bytes[i]);
-  }
-  base64.flush();
-  return "data:image/gif;base64," + base64;
-};
-qrcode.stringToBytes;
-function _extends() {
-  return _extends = Object.assign ? Object.assign.bind() : function(n) {
-    for (var e = 1; e < arguments.length; e++) {
-      var t = arguments[e];
-      for (var r in t) ({}).hasOwnProperty.call(t, r) && (n[r] = t[r]);
+  };
+  QRCode2.PAD0 = 236;
+  QRCode2.PAD1 = 17;
+  QRCode2.createData = function(typeNumber, errorCorrectLevel, dataList) {
+    var rsBlocks = RSBlock2.getRSBlocks(typeNumber, errorCorrectLevel);
+    var buffer = new BitBuffer2();
+    for (var i = 0; i < dataList.length; i++) {
+      var data = dataList[i];
+      buffer.put(data.mode, 4);
+      buffer.put(data.getLength(), util2.getLengthInBits(data.mode, typeNumber));
+      data.write(buffer);
     }
-    return n;
-  }, _extends.apply(null, arguments);
+    var totalDataCount = 0;
+    for (var i = 0; i < rsBlocks.length; i++) {
+      totalDataCount += rsBlocks[i].dataCount;
+    }
+    if (buffer.getLengthInBits() > totalDataCount * 8) {
+      throw new Error("code length overflow. (" + buffer.getLengthInBits() + ">" + totalDataCount * 8 + ")");
+    }
+    if (buffer.getLengthInBits() + 4 <= totalDataCount * 8) {
+      buffer.put(0, 4);
+    }
+    while (buffer.getLengthInBits() % 8 != 0) {
+      buffer.putBit(false);
+    }
+    while (true) {
+      if (buffer.getLengthInBits() >= totalDataCount * 8) {
+        break;
+      }
+      buffer.put(QRCode2.PAD0, 8);
+      if (buffer.getLengthInBits() >= totalDataCount * 8) {
+        break;
+      }
+      buffer.put(QRCode2.PAD1, 8);
+    }
+    return QRCode2.createBytes(buffer, rsBlocks);
+  };
+  QRCode2.createBytes = function(buffer, rsBlocks) {
+    var offset = 0;
+    var maxDcCount = 0;
+    var maxEcCount = 0;
+    var dcdata = new Array(rsBlocks.length);
+    var ecdata = new Array(rsBlocks.length);
+    for (var r = 0; r < rsBlocks.length; r++) {
+      var dcCount = rsBlocks[r].dataCount;
+      var ecCount = rsBlocks[r].totalCount - dcCount;
+      maxDcCount = Math.max(maxDcCount, dcCount);
+      maxEcCount = Math.max(maxEcCount, ecCount);
+      dcdata[r] = new Array(dcCount);
+      for (var i = 0; i < dcdata[r].length; i++) {
+        dcdata[r][i] = 255 & buffer.buffer[i + offset];
+      }
+      offset += dcCount;
+      var rsPoly = util2.getErrorCorrectPolynomial(ecCount);
+      var rawPoly = new Polynomial2(dcdata[r], rsPoly.getLength() - 1);
+      var modPoly = rawPoly.mod(rsPoly);
+      ecdata[r] = new Array(rsPoly.getLength() - 1);
+      for (var i = 0; i < ecdata[r].length; i++) {
+        var modIndex = i + modPoly.getLength() - ecdata[r].length;
+        ecdata[r][i] = modIndex >= 0 ? modPoly.get(modIndex) : 0;
+      }
+    }
+    var totalCodeCount = 0;
+    for (var i = 0; i < rsBlocks.length; i++) {
+      totalCodeCount += rsBlocks[i].totalCount;
+    }
+    var data = new Array(totalCodeCount);
+    var index = 0;
+    for (var i = 0; i < maxDcCount; i++) {
+      for (var r = 0; r < rsBlocks.length; r++) {
+        if (i < dcdata[r].length) {
+          data[index++] = dcdata[r][i];
+        }
+      }
+    }
+    for (var i = 0; i < maxEcCount; i++) {
+      for (var r = 0; r < rsBlocks.length; r++) {
+        if (i < ecdata[r].length) {
+          data[index++] = ecdata[r][i];
+        }
+      }
+    }
+    return data;
+  };
+  QRCode_1 = QRCode2;
+  return QRCode_1;
 }
-function _objectWithoutProperties(e, t) {
-  if (null == e) return {};
-  var o, r, i = _objectWithoutPropertiesLoose(e, t);
-  if (Object.getOwnPropertySymbols) {
-    var n = Object.getOwnPropertySymbols(e);
-    for (r = 0; r < n.length; r++) o = n[r], -1 === t.indexOf(o) && {}.propertyIsEnumerable.call(e, o) && (i[o] = e[o]);
-  }
-  return i;
-}
-function _objectWithoutPropertiesLoose(r, e) {
-  if (null == r) return {};
-  var t = {};
-  for (var n in r) if ({}.hasOwnProperty.call(r, n)) {
-    if (-1 !== e.indexOf(n)) continue;
-    t[n] = r[n];
-  }
-  return t;
-}
-var _excluded$1 = ["bgColor", "bgD", "fgD", "fgColor", "size", "title", "viewBoxSize", "xmlns"];
-var propTypes$1 = {
-  bgColor: PropTypes.oneOfType([PropTypes.object, PropTypes.string]).isRequired,
-  bgD: PropTypes.string.isRequired,
-  fgColor: PropTypes.oneOfType([PropTypes.object, PropTypes.string]).isRequired,
-  fgD: PropTypes.string.isRequired,
-  size: PropTypes.number.isRequired,
-  title: PropTypes.string,
-  viewBoxSize: PropTypes.number.isRequired,
-  xmlns: PropTypes.string
-};
-var QRCodeSvg = /* @__PURE__ */ reactExports.forwardRef(function(_ref, ref) {
-  var bgColor = _ref.bgColor, bgD = _ref.bgD, fgD = _ref.fgD, fgColor = _ref.fgColor, size = _ref.size, title = _ref.title, viewBoxSize = _ref.viewBoxSize, _ref$xmlns = _ref.xmlns, xmlns = _ref$xmlns === void 0 ? "http://www.w3.org/2000/svg" : _ref$xmlns, props = _objectWithoutProperties(_ref, _excluded$1);
-  return /* @__PURE__ */ React.createElement("svg", _extends({}, props, {
-    height: size,
-    ref,
-    viewBox: "0 0 ".concat(viewBoxSize, " ").concat(viewBoxSize),
-    width: size,
-    xmlns
-  }), title ? /* @__PURE__ */ React.createElement("title", null, title) : null, /* @__PURE__ */ React.createElement("path", {
-    d: bgD,
-    fill: bgColor
-  }), /* @__PURE__ */ React.createElement("path", {
-    d: fgD,
-    fill: fgColor
-  }));
-});
-QRCodeSvg.displayName = "QRCodeSvg";
-QRCodeSvg.propTypes = propTypes$1;
-var _excluded = ["bgColor", "fgColor", "level", "size", "value"];
-qrcode.stringToBytes = function(s) {
-  return Array.from(new TextEncoder().encode(s));
-};
-var propTypes = {
-  bgColor: PropTypes.oneOfType([PropTypes.object, PropTypes.string]),
-  fgColor: PropTypes.oneOfType([PropTypes.object, PropTypes.string]),
-  level: PropTypes.string,
-  size: PropTypes.number,
-  value: PropTypes.string.isRequired
-};
-var QRCode = /* @__PURE__ */ reactExports.forwardRef(function(_ref, ref) {
-  var _ref$bgColor = _ref.bgColor, bgColor = _ref$bgColor === void 0 ? "#FFFFFF" : _ref$bgColor, _ref$fgColor = _ref.fgColor, fgColor = _ref$fgColor === void 0 ? "#000000" : _ref$fgColor, _ref$level = _ref.level, level = _ref$level === void 0 ? "L" : _ref$level, _ref$size = _ref.size, size = _ref$size === void 0 ? 256 : _ref$size, value = _ref.value, props = _objectWithoutProperties(_ref, _excluded);
-  var qr = qrcode(0, level);
-  qr.addData(value);
-  qr.make();
-  var moduleCount = qr.getModuleCount();
-  var cells = Array.from({
-    length: moduleCount
-  }, function(_, rowIndex) {
-    return Array.from({
-      length: moduleCount
-    }, function(_2, colIndex) {
-      return qr.isDark(rowIndex, colIndex);
-    });
+var QRCodeSvg = {};
+var hasRequiredQRCodeSvg;
+function requireQRCodeSvg() {
+  if (hasRequiredQRCodeSvg) return QRCodeSvg;
+  hasRequiredQRCodeSvg = 1;
+  Object.defineProperty(QRCodeSvg, "__esModule", {
+    value: true
   });
-  return /* @__PURE__ */ React.createElement(QRCodeSvg, _extends({}, props, {
-    bgColor,
-    bgD: cells.map(function(row, rowIndex) {
-      return row.map(function(cell, cellIndex) {
-        return !cell ? "M ".concat(cellIndex, " ").concat(rowIndex, " l 1 0 0 1 -1 0 Z") : "";
-      }).join(" ");
-    }).join(" "),
-    fgColor,
-    fgD: cells.map(function(row, rowIndex) {
-      return row.map(function(cell, cellIndex) {
-        return cell ? "M ".concat(cellIndex, " ").concat(rowIndex, " l 1 0 0 1 -1 0 Z") : "";
-      }).join(" ");
-    }).join(" "),
-    ref,
-    size,
-    viewBoxSize: moduleCount
-  }));
-});
-QRCode.displayName = "QRCode";
-QRCode.propTypes = propTypes;
+  var _extends = Object.assign || function(target) {
+    for (var i = 1; i < arguments.length; i++) {
+      var source = arguments[i];
+      for (var key in source) {
+        if (Object.prototype.hasOwnProperty.call(source, key)) {
+          target[key] = source[key];
+        }
+      }
+    }
+    return target;
+  };
+  var _propTypes = /* @__PURE__ */ requirePropTypes();
+  var _propTypes2 = _interopRequireDefault(_propTypes);
+  var _react = requireReact();
+  var _react2 = _interopRequireDefault(_react);
+  function _interopRequireDefault(obj) {
+    return obj && obj.__esModule ? obj : { default: obj };
+  }
+  function _objectWithoutProperties(obj, keys) {
+    var target = {};
+    for (var i in obj) {
+      if (keys.indexOf(i) >= 0) continue;
+      if (!Object.prototype.hasOwnProperty.call(obj, i)) continue;
+      target[i] = obj[i];
+    }
+    return target;
+  }
+  var propTypes2 = {
+    bgColor: _propTypes2.default.oneOfType([_propTypes2.default.object, _propTypes2.default.string]).isRequired,
+    bgD: _propTypes2.default.string.isRequired,
+    fgColor: _propTypes2.default.oneOfType([_propTypes2.default.object, _propTypes2.default.string]).isRequired,
+    fgD: _propTypes2.default.string.isRequired,
+    size: _propTypes2.default.number.isRequired,
+    title: _propTypes2.default.string,
+    viewBoxSize: _propTypes2.default.number.isRequired,
+    xmlns: _propTypes2.default.string
+  };
+  var QRCodeSvg$1 = (0, _react.forwardRef)(function(_ref, ref) {
+    var bgColor = _ref.bgColor, bgD = _ref.bgD, fgD = _ref.fgD, fgColor = _ref.fgColor, size = _ref.size, title = _ref.title, viewBoxSize = _ref.viewBoxSize, _ref$xmlns = _ref.xmlns, xmlns = _ref$xmlns === void 0 ? "http://www.w3.org/2000/svg" : _ref$xmlns, props = _objectWithoutProperties(_ref, ["bgColor", "bgD", "fgD", "fgColor", "size", "title", "viewBoxSize", "xmlns"]);
+    return _react2.default.createElement(
+      "svg",
+      _extends({}, props, { height: size, ref, viewBox: "0 0 " + viewBoxSize + " " + viewBoxSize, width: size, xmlns }),
+      title ? _react2.default.createElement(
+        "title",
+        null,
+        title
+      ) : null,
+      _react2.default.createElement("path", { d: bgD, fill: bgColor }),
+      _react2.default.createElement("path", { d: fgD, fill: fgColor })
+    );
+  });
+  QRCodeSvg$1.displayName = "QRCodeSvg";
+  QRCodeSvg$1.propTypes = propTypes2;
+  QRCodeSvg.default = QRCodeSvg$1;
+  return QRCodeSvg;
+}
+var hasRequiredLib;
+function requireLib() {
+  if (hasRequiredLib) return lib;
+  hasRequiredLib = 1;
+  Object.defineProperty(lib, "__esModule", {
+    value: true
+  });
+  lib.QRCode = void 0;
+  var _extends = Object.assign || function(target) {
+    for (var i = 1; i < arguments.length; i++) {
+      var source = arguments[i];
+      for (var key in source) {
+        if (Object.prototype.hasOwnProperty.call(source, key)) {
+          target[key] = source[key];
+        }
+      }
+    }
+    return target;
+  };
+  var _propTypes = /* @__PURE__ */ requirePropTypes();
+  var _propTypes2 = _interopRequireDefault(_propTypes);
+  var _ErrorCorrectLevel = requireErrorCorrectLevel();
+  var _ErrorCorrectLevel2 = _interopRequireDefault(_ErrorCorrectLevel);
+  var _QRCode = requireQRCode();
+  var _QRCode2 = _interopRequireDefault(_QRCode);
+  var _react = requireReact();
+  var _react2 = _interopRequireDefault(_react);
+  var _QRCodeSvg = requireQRCodeSvg();
+  var _QRCodeSvg2 = _interopRequireDefault(_QRCodeSvg);
+  function _interopRequireDefault(obj) {
+    return obj && obj.__esModule ? obj : { default: obj };
+  }
+  function _objectWithoutProperties(obj, keys) {
+    var target = {};
+    for (var i in obj) {
+      if (keys.indexOf(i) >= 0) continue;
+      if (!Object.prototype.hasOwnProperty.call(obj, i)) continue;
+      target[i] = obj[i];
+    }
+    return target;
+  }
+  function bytesToBinaryString(bytes) {
+    return bytes.map(function(b) {
+      return String.fromCharCode(b & 255);
+    }).join("");
+  }
+  function encodeStringToUtf8Bytes(input) {
+    return Array.from(new TextEncoder().encode(input));
+  }
+  var propTypes2 = {
+    bgColor: _propTypes2.default.oneOfType([_propTypes2.default.object, _propTypes2.default.string]),
+    fgColor: _propTypes2.default.oneOfType([_propTypes2.default.object, _propTypes2.default.string]),
+    level: _propTypes2.default.string,
+    size: _propTypes2.default.number,
+    value: _propTypes2.default.string.isRequired
+  };
+  var QRCode2 = (0, _react.forwardRef)(function(_ref, ref) {
+    var _ref$bgColor = _ref.bgColor, bgColor = _ref$bgColor === void 0 ? "#FFFFFF" : _ref$bgColor, _ref$fgColor = _ref.fgColor, fgColor = _ref$fgColor === void 0 ? "#000000" : _ref$fgColor, _ref$level = _ref.level, level = _ref$level === void 0 ? "L" : _ref$level, _ref$size = _ref.size, size = _ref$size === void 0 ? 256 : _ref$size, value = _ref.value, props = _objectWithoutProperties(_ref, ["bgColor", "fgColor", "level", "size", "value"]);
+    var qrcode = new _QRCode2.default(-1, _ErrorCorrectLevel2.default[level]);
+    var utf8Bytes = encodeStringToUtf8Bytes(value);
+    var binaryString = bytesToBinaryString(utf8Bytes);
+    qrcode.addData(binaryString, "Byte");
+    qrcode.make();
+    var cells = qrcode.modules;
+    return _react2.default.createElement(_QRCodeSvg2.default, _extends({}, props, {
+      bgColor,
+      bgD: cells.map(function(row, rowIndex) {
+        return row.map(function(cell, cellIndex) {
+          return !cell ? "M " + cellIndex + " " + rowIndex + " l 1 0 0 1 -1 0 Z" : "";
+        }).join(" ");
+      }).join(" "),
+      fgColor,
+      fgD: cells.map(function(row, rowIndex) {
+        return row.map(function(cell, cellIndex) {
+          return cell ? "M " + cellIndex + " " + rowIndex + " l 1 0 0 1 -1 0 Z" : "";
+        }).join(" ");
+      }).join(" "),
+      ref,
+      size,
+      viewBoxSize: cells.length
+    }));
+  });
+  lib.QRCode = QRCode2;
+  QRCode2.displayName = "QRCode";
+  QRCode2.propTypes = propTypes2;
+  lib.default = QRCode2;
+  return lib;
+}
+var libExports = requireLib();
+const QRCode = /* @__PURE__ */ getDefaultExportFromCjs(libExports);
 function buildTotpQrImageOptions() {
   return {
     bgColor: "#ffffff",
@@ -1949,15 +1346,15 @@ function resolveSettingsSaveProofModalCopy(input) {
     confirmLabel: "Save settings"
   };
 }
-function isSettingsSaveProofSubmitDisabled(mode, credentials, totpRequired) {
+function isSettingsSaveProofSubmitDisabled(mode2, credentials, totpRequired) {
   const current = credentials.currentPassword?.trim() ?? "";
   const next = credentials.newPassword?.trim() ?? "";
   const confirm = credentials.confirmPassword?.trim() ?? "";
   const totp = credentials.totpCode?.trim() ?? "";
-  if (mode === "setup-gate") {
+  if (mode2 === "setup-gate") {
     return next.length === 0 || confirm.length === 0 || next !== confirm;
   }
-  if (mode === "change-password") {
+  if (mode2 === "change-password") {
     if (current.length === 0 || next.length === 0 || confirm.length === 0 || next !== confirm) {
       return true;
     }
@@ -2587,8 +1984,8 @@ function normalizeGuardSettings(settings) {
 }
 function buildConsequenceSummary(settings) {
   const level = settings.security_level;
-  const mode = settings.mode;
-  if (mode === "observe") return "Guard is watching and recording what your AI apps do, but it will not pause any actions. Switch to Prompt or Enforce when you want Guard to actively protect you.";
+  const mode2 = settings.mode;
+  if (mode2 === "observe") return "Guard is watching and recording what your AI apps do, but it will not pause any actions. Switch to Prompt or Enforce when you want Guard to actively protect you.";
   if (level === "relaxed") return "Guard will warn about destructive commands and credential sharing but will not pause for approval. Most safe actions run automatically. Good for trusted environments.";
   if (level === "balanced") return "Guard will ask before secret access, hidden execution, and destructive commands. New network destinations get a warning. This is the recommended setting for most users.";
   if (level === "strict") return "Guard will ask before almost every risky action, including new network destinations. Use this when working with sensitive data or untrusted AI tools.";
@@ -2617,18 +2014,18 @@ function applyApprovalGateDraft(settings, updates) {
     }
   };
 }
-function protectionModeHelp(mode) {
-  if (mode === "enforce") {
+function protectionModeHelp(mode2) {
+  if (mode2 === "enforce") {
     return "Guard keeps risky actions stopped until you allow them.";
   }
-  if (mode === "observe") {
+  if (mode2 === "observe") {
     return "Guard logs what it sees without pausing anything.";
   }
   return "Guard pauses risky actions and asks what to do.";
 }
-function protectionModeLabel(mode) {
-  const match = protectionModeChoices.find((choice) => choice.value === mode);
-  return match?.label ?? mode;
+function protectionModeLabel(mode2) {
+  const match = protectionModeChoices.find((choice) => choice.value === mode2);
+  return match?.label ?? mode2;
 }
 function saveStatusText(saveSuccess, saveError) {
   if (saveSuccess) {
@@ -2892,8 +2289,8 @@ function SettingsWorkspace({ onApprovalGateChange }) {
       onApprovalGateChange?.(gate);
     }
   }, [onApprovalGateChange]);
-  const openProofModal = reactExports.useCallback((mode, action) => {
-    setProofModalMode(mode);
+  const openProofModal = reactExports.useCallback((mode2, action) => {
+    setProofModalMode(mode2);
     setPendingProofAction(action);
     setProofModalError(null);
     setProofModalOpen(true);
