+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/supply-chain-protection-stats.js
@@ -1,4 +1,4 @@
-import { au as GuardHarnessActionError, j as jsxRuntimeExports, r as reactExports, N as HiMiniKey, A as ActionButton, S as SectionLabel, Y as fetchSettings } from "../guard-dashboard.js";
+import { au as GuardHarnessActionError, j as jsxRuntimeExports, r as reactExports, O as HiMiniKey, A as ActionButton, S as SectionLabel, Y as fetchSettings } from "../guard-dashboard.js";
 const APPROVAL_GATE_REQUIRED_CODES = /* @__PURE__ */ new Set([
   "approval_gate_required",
   "approval_gate_password_required",
