+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/supply-chain-workspace.js
@@ -1,4 +1,4 @@
-import { aG as isSupplyChainAuditIncomplete, aH as readString$1, aI as isRecord$2, au as GuardHarnessActionError, r as reactExports, j as jsxRuntimeExports, d as HiMiniCheckCircle, aw as HiMiniArrowPath, w as HiMiniExclamationTriangle, ac as Tag, m as formatRelativeTime, aJ as HiMiniClock, aK as IconActionButton, I as HiMiniXCircle, ax as HiMiniTrash, l as HiMiniShieldCheck, F as HiMiniWrenchScrewdriver, aL as HiMiniBeaker, aM as ActivationSummary, aN as ActionResultPanel, ad as HiMiniMagnifyingGlass, b as EmptyState, A as ActionButton, aO as HiMiniBugAnt, o as HiMiniXMark, aP as GuardModalLayer, aQ as ConnectFlowCard, aE as HiMiniArrowTopRightOnSquare, aR as HiMiniCloudArrowDown, aS as fetchPackageFirewallStatus, aT as runPackageAudit, aU as resolveSupplyChainAuditFailure, aV as runPackageSync, aW as startPackageFirewallConnect, aX as runPackageFirewallAction, aY as parseInterceptProofSnapshot, aZ as openPackageFirewallShell, S as SectionLabel, a_ as EntitlementNotice, a$ as fetchSupplyChainBundle, B as Badge, b0 as HiMiniDocumentMagnifyingGlass, b1 as HiMiniShieldExclamation, b2 as HiMiniComputerDesktop, t as HiMiniCloud, b3 as HiMiniArrowDown, b4 as HiMiniArrowUp, ah as HiMiniArrowLeft, b5 as HiMiniArrowRight, b6 as HiMiniCloudArrowUp, b7 as HiMiniInformationCircle, b8 as fetchReceipts, h as harnessDisplayName, p as HiMiniChevronUp, q as HiMiniChevronDown } from "../guard-dashboard.js";
+import { aQ as isSupplyChainAuditIncomplete, aR as readString$1, aS as isRecord$2, au as GuardHarnessActionError, r as reactExports, j as jsxRuntimeExports, d as HiMiniCheckCircle, aw as HiMiniArrowPath, w as HiMiniExclamationTriangle, ac as Tag, m as formatRelativeTime, aT as HiMiniClock, aU as IconActionButton, I as HiMiniXCircle, ax as HiMiniTrash, l as HiMiniShieldCheck, F as HiMiniWrenchScrewdriver, aV as HiMiniBeaker, aW as ActivationSummary, aX as ActionResultPanel, ad as HiMiniMagnifyingGlass, b as EmptyState, A as ActionButton, aY as HiMiniBugAnt, o as HiMiniXMark, aZ as GuardModalLayer, a_ as ConnectFlowCard, aO as HiMiniArrowTopRightOnSquare, a$ as HiMiniCloudArrowDown, b0 as fetchPackageFirewallStatus, b1 as runPackageAudit, b2 as resolveSupplyChainAuditFailure, b3 as runPackageSync, b4 as startPackageFirewallConnect, b5 as runPackageFirewallAction, b6 as parseInterceptProofSnapshot, b7 as openPackageFirewallShell, S as SectionLabel, b8 as EntitlementNotice, b9 as fetchSupplyChainBundle, B as Badge, ba as HiMiniDocumentMagnifyingGlass, bb as HiMiniShieldExclamation, bc as HiMiniComputerDesktop, t as HiMiniCloud, bd as HiMiniArrowDown, be as HiMiniArrowUp, ah as HiMiniArrowLeft, bf as HiMiniArrowRight, aE as HiMiniCloudArrowUp, bg as HiMiniInformationCircle, bh as fetchReceipts, h as harnessDisplayName, p as HiMiniChevronUp, q as HiMiniChevronDown } from "../guard-dashboard.js";
 import { i as isGuardHarnessActionError, a as isSupplyChainSyncConnectError, r as readHarnessActionErrorCode, A as ApprovalProofInline, u as useResolvedApprovalGate, b as resolveApprovalGateSyncFailure, c as isApprovalGateRequiredError, d as readHarnessActionUserMessage, e as ApprovalProofModal, f as buildSupplyChainStats } from "./supply-chain-protection-stats.js";
 import { resolveFeedStaleness } from "./feed-health-workspace.js";
 import { r as resolveHomeProtectionStatus } from "./home-protection-module.js";
